fix rsc error & update locales & update method to detect ai supported books
