+++ locales/bn/common.json
@@ -2,6 +2,7 @@
   "advanced-search": "উন্নত অনুসন্ধান",
   "load-more": "আরও লোড করুন",
   "clear-all": "সব মুছে দিন",
+  "ai": "এআই",
   "search": "অনুসন্ধান",
   "pagination": {
     "previous": "পূর্ববর্তী",
