+++ locales/es/common.json
@@ -2,6 +2,7 @@
   "advanced-search": "Búsqueda Avanzada",
   "load-more": "Cargar Más",
   "clear-all": "Borrar todo",
+  "ai": "IA",
   "search": "Buscar",
   "pagination": {
     "previous": "Anterior",
