+++ locales/fa/common.json
@@ -2,6 +2,7 @@
   "advanced-search": "جستجوی پیشرفته",
   "load-more": "بارگذاری بیشتر",
   "clear-all": "پاک کردن همه",
+  "ai": "هوش مصنوعی",
   "search": "جستجو",
   "pagination": {
     "previous": "قبلی",
