+++ locales/fr/common.json
@@ -2,6 +2,7 @@
   "advanced-search": "Recherche Avancée",
   "load-more": "Charger plus",
   "clear-all": "Effacer tout",
+  "ai": "IA",
   "search": "Chercher",
   "pagination": {
     "previous": "Précédent",
