+++ locales/ha/common.json
@@ -2,6 +2,7 @@
   "advanced-search": "Bincike Mai Zurfi",
   "load-more": "Daukar Karin",
   "clear-all": "Goge duk",
+  "ai": "AI",
   "search": "Bincike",
   "pagination": {
     "previous": "Na Baya",
