+++ locales/hi/common.json
@@ -2,6 +2,7 @@
   "advanced-search": "उन्नत खोज",
   "load-more": "और लोड करें",
   "clear-all": "सभी साफ़ करें",
+  "ai": "एआई",
   "search": "खोज",
   "pagination": {
     "previous": "पिछला",
