+++ locales/ms/common.json
@@ -2,6 +2,7 @@
   "advanced-search": "Carian Lanjutan",
   "load-more": "Muat Lebih",
   "clear-all": "Kosongkan semua",
+  "ai": "AI",
   "search": "Cari",
   "pagination": {
     "previous": "Sebelumnya",
