+++ locales/ps/common.json
@@ -2,6 +2,7 @@
   "advanced-search": "لټون پرمختللی",
   "load-more": "نور ولولې",
   "clear-all": "ټول پاک کړئ",
+  "ai": "AI",
   "search": "لټون",
   "pagination": {
     "previous": "مخکې",
