+++ locales/ru/common.json
@@ -2,6 +2,7 @@
   "advanced-search": "Расширеннный поиск",
   "load-more": "Загрузить больше",
   "clear-all": "Очистить все",
+  "ai": "ИИ",
   "search": "Искать",
   "pagination": {
     "previous": "Предыдущее",
