+++ locales/ru/reader.json
@@ -13,6 +13,28 @@
     "input-placeholder": "Номер страницы",
     "go": "Перейти"
   },
+  "search": {
+    "begin-search": "Начните поиск",
+    "description": "Искусственный интеллект поиск находит самые близкие совпадения, даже если вы не знаете точную фразу.",
+    "view-in-advanced": "Просмотреть в Расширенном поиске",
+    "best-match": "Лучшее совпадение",
+    "order": "Заказ (скоро)",
+    "portion": "Часть",
+    "entire-book": "Вся книга",
+    "book": "Книга",
+    "book-select-description": "Чтобы искать по нескольким книгам, используйте Расширенный поиск"
+  },
+  "chat": {
+    "new-chat": "Новый чат",
+    "share-chat": "Поделиться чатом",
+    "copy": "Копировать в буфер обмена",
+    "copied": "Скопировано в буфер обмена!",
+    "regenerate": "Пересоздать",
+    "mark-as-correct": "Отметить как правильное",
+    "report-as-incorrect": "Отметить как неправильное",
+    "sources": "Источники",
+    "pg-x": "Стр. {page, number, nogroup}"
+  },
   "version": "Версия",
   "select-version": "Выберите версию",
   "other-titles": "Другие названия",
