+++ locales/so/common.json
@@ -2,6 +2,7 @@
   "advanced-search": "Raadin Horumarsan",
   "load-more": "Ku Shub wax dheeraad ah",
   "clear-all": "Nadiifi dhammaan",
+  "ai": "AI",
   "search": "Raadi",
   "pagination": {
     "previous": "Kii Hore",
