+++ locales/tr/common.json
@@ -2,6 +2,7 @@
   "advanced-search": "Gelişmiş Arama",
   "load-more": "Daha Fazla Yükle",
   "clear-all": "Tümünü Temizle",
+  "ai": "AI",
   "search": "Arama",
   "pagination": {
     "previous": "Önceki",
