+++ locales/tr/reader.json
@@ -13,6 +13,28 @@
     "input-placeholder": "Sayfa numarası",
     "go": "Git"
   },
+  "search": {
+    "begin-search": "Aramaya başlayın",
+    "description": " Yapay Zeka Arama, tam ifadeyi bilmiyorsanız bile en yakın eşleşmeleri bulur.",
+    "view-in-advanced": "Gelişmiş Arama'da Görüntüle",
+    "best-match": "En İyi Eşleşme",
+    "order": "Sırala (yakında)",
+    "portion": "Bölüm",
+    "entire-book": "Tüm Kitap",
+    "book": "Kitap",
+    "book-select-description": "Birden fazla kitap aramak için Gelişmiş Arama'yı kullanın"
+  },
+  "chat": {
+    "new-chat": "Yeni Sohbet",
+    "share-chat": "Sohbeti Paylaş",
+    "copy": "Panoya Kopyala",
+    "copied": "Panoya kopyalandı!",
+    "regenerate": "Yeniden oluştur",
+    "mark-as-correct": "Doğru olarak işaretle",
+    "report-as-incorrect": "Yanlış olarak bildir",
+    "sources": "Kaynaklar",
+    "pg-x": "Sf {page, number, nogroup}"
+  },
   "version": "Versiyon",
   "select-version": "Bir versiyon seçin",
   "other-titles": "Diğer başlıklar",
