+++ locales/ur/common.json
@@ -2,6 +2,7 @@
   "advanced-search": "اعلیٰ تلاش",
   "load-more": "مزید لوڈ کریں",
   "clear-all": "سب صاف کریں",
+  "ai": "اے آئی",
   "search": "تلاش کریں",
   "pagination": {
     "previous": "پچھلا",
