+++ src/app/[locale]/t/[bookId]/_components/tab-content.tsx
@@ -1,3325 +1,8 @@
 "use client";
 
-import { useLocale, useTranslations } from "next-intl";
-import type { AppLocale } from "~/i18n.config";
-import SidebarContainer from "./sidebar/sidebar-container";
-import { Alert, AlertDescription, AlertTitle } from "@/components/ui/alert";
-import { getLocaleDirection } from "@/lib/locale/utils";
-import type { TabProps, tabs } from "./sidebar/tabs";
-import dynamic from "next/dynamic";
-import Spinner from "@/components/ui/spinner";
-
-const supportedBooks = {
-  sahih: true,
-  "diwan-27": true,
-  "diwan-35": true,
-  "diwan-39": true,
-  "diwan-40": true,
-  "diwan-41": true,
-  "gharib-quran": true,
-  "diwan-44": true,
-  "fadail-makka": true,
-  "diwan-55": true,
-  "nasikh-wa-mansukh": true,
-  "tanzil-quran": true,
-  "ahadith-yazid": true,
-  "nuskhat-zubayr": true,
-  "ahadith-min-maghazi": true,
-  "adab-murua": true,
-  manasik: true,
-  faraid: true,
-  hadith: true,
-  mufaddaliyyat: true,
-  "jamharat-ashcar-carab": true,
-  cayn: true,
-  "casharat-ahadith": true,
-  cawali: true,
-  "majlis-min-fawaid": true,
-  "kitab-sibawayhi": true,
-  jihad: true,
-  "ikhtilaf-abi-hanifa": true,
-  shicr: true,
-  "makharij-fi-hiyal": true,
-  "majalis-ibn-qasim": true,
-  duca: true,
-  amthal: true,
-  "hadhf-min-nasab-quraysh": true,
-  "muharaba-min-muwatta-ibn-wahb": true,
-  "qada-min-muwatta-ibn-wahb": true,
-  qadar: true,
-  "tafsir-quran": true,
-  "juz-fihi-hadith": true,
-  "juz-hadith": true,
-  fitna: true,
-  "musnad-4": true,
-  "nasab-khayl": true,
-  "nasab-macad": true,
-  "ikhtilaf-hadith": true,
-  tafsir: true,
-  "muntakhab-kalam-fi-tafsir-ahlam": true,
-  "fiqh-absat": true,
-  "fiqh-akbar": true,
-  "musnad-1": true,
-  "tafsir-muqatil": true,
-  "tafsir-2": true,
-  mashyakha: true,
-  "amthal-carab": true,
-  "jumal-fi-nahw": true,
-  muwatta: true,
-  "hadith-ismacil": true,
-  "musnad-2": true,
-  zuhd: true,
-  kharaj: true,
-  "radd-cala-siyar-awzaci": true,
-  "zuhd-1": true,
-  siyar: true,
-  "mushtabihat-quran": true,
-  "athar-1": true,
-  "hujja-cala-ahl-madina": true,
-  "siyar-saghir": true,
-  "nuskhat-wakic": true,
-  "zuhd-2": true,
-  "kharaj-1": true,
-  asnam: true,
-  "jimac-cilm": true,
-  "musnad-5": true,
-  risala: true,
-  "sunan-mathura": true,
-  "tafsir-shafici": true,
-  umm: true,
-  jim: true,
-  azmina: true,
-  "juz-1": true,
-  "lughat-quran": true,
-  "macani-quran": true,
-  "futuh-sham": true,
-  ridda: true,
-  "juz-ashyab": true,
-  dibaj: true,
-  khayl: true,
-  "majaz-quran": true,
-  "sharh-naqaid-jarir": true,
-  "amali-fi-athar": true,
-  musannaf: true,
-  "tafsir-4": true,
-  "zuhd-3": true,
-  "sirat-cumar-ibn-cabd-caziz": true,
-  "nawadir-fi-lugha": true,
-  "macani-quran-1": true,
-  qawafi: true,
-  "hadith-ansari": true,
-  asmaciyyat: true,
-  "fuhulat-shucara": true,
-  ibl: true,
-  "khalq-insan": true,
-  sha: true,
-  nuskha: true,
-  "musnad-6": true,
-  "juz-2": true,
-  "akhbar-wafidat": true,
-  "akhbar-wafidin": true,
-  "hadith-2": true,
-  "nuskha-1": true,
-  "amthal-1": true,
-  amwal: true,
-  "fadail-quran": true,
-  "gharib-hadith": true,
-  "gharib-musannaf": true,
-  "khutab-wa-mawaciz": true,
-  "lughat-qabail": true,
-  "nasikh-wa-mansukh-1": true,
-  sunan: true,
-  "tafsir-min-sunan-sacid": true,
-  "juz-3": true,
-  fitan: true,
-  nawadir: true,
-  "musnad-7": true,
-  ashriba: true,
-  "tabaqat-kubra": true,
-  "juz-4": true,
-  "sharh-diwan-dhi-rumma": true,
-  wahshiyyat: true,
-  "hadith-3": true,
-  "waqf-wa-ibtida": true,
-  "hadith-yahya": true,
-  "min-kalam-fi-rijal": true,
-  "sualat-ibn-junayd": true,
-  "tasmiyat-man-ruwiya": true,
-  cilm: true,
-  cilal: true,
-  "sualat-ibn-abi-shayba": true,
-  adab: true,
-  "musannaf-1": true,
-  "musnad-8": true,
-  qada: true,
-  "nasab-quraysh": true,
-  "ashrat-saca": true,
-  "mukhtasar-fi-tibb": true,
-  "musnad-9": true,
-  "karam-wa-jawd": true,
-  "diwan-77": true,
-  "musnad-10": true,
-  tabaqat: true,
-  tarikh: true,
-  "asami-wa-kuna": true,
-  tawahhum: true,
-  "fadail-sahaba": true,
-  "masail-riwayat-cabd-allah": true,
-  "masail-riwayat-ibn-hani": true,
-  "masail-riwayat-salih": true,
-  "radd-cala-zanadiqa": true,
-  "sualat-abi-dawud-li-ibn-hanbal": true,
-  "usul-sunna-1": true,
-  "zuhd-4": true,
-  arbacun: true,
-  "zuhd-5": true,
-  "adab-nufus": true,
-  "fahm-quran": true,
-  "maiyyat-caql": true,
-  makasib: true,
-  "risalat-mustarshidin": true,
-  "iman-2": true,
-  alfaz: true,
-  "islah-mantiq": true,
-  "kanz-lughawi": true,
-  "qalb-wa-ibdal": true,
-  "juz-hadith-1": true,
-  "man-nusiba-li-ummihi": true,
-  muhabbar: true,
-  "mukhtalaf-qabail": true,
-  "munammaq-fi-akhbar-quraysh": true,
-  "musnad-sacd-ibn-abi-waqqas": true,
-  "wasaya-muluk": true,
-  "birr-wa-sila": true,
-  "min-cawali-musnad": true,
-  "muntakhab-min-musnad": true,
-  "masail-3": true,
-  "amwal-1": true,
-  "juz-muammal": true,
-  "jadd-wa-hazl": true,
-  "sunan-1": true,
-  "balagha-wa-ijaz": true,
-  "bayan-wa-tabyin": true,
-  bukhala: true,
-  "cadawa-wa-hasad": true,
-  cuthmaniyya: true,
-  "dhamm-akhlaq-kuttab": true,
-  "fadl-hashim": true,
-  "fakhr-sudan": true,
-  futya: true,
-  "hakamayn-wa-taswib-cali": true,
-  "hasid-wa-mahsud": true,
-  hayawan: true,
-  "istinjaz-wacd": true,
-  "kitman-sirr": true,
-  "macad-wa-macash": true,
-  "madh-nabidh": true,
-  "madh-tujjar": true,
-  "mahasin-wa-addad": true,
-  "manaqib-turk": true,
-  "mawadda-wa-khulta": true,
-  mucallimin: true,
-  "mufakharat-jawari": true,
-  nubl: true,
-  "risala-cabbasiyya": true,
-  "risala-ila-abi-faraj": true,
-  "sharib-wa-mashrub": true,
-  "sinacat-quwwad": true,
-  "tabaqat-mughannin": true,
-  "tafdil-batn": true,
-  "tafdil-nutq": true,
-  "tarbic-wa-tadwir": true,
-  wukala: true,
-  "adab-mufrad": true,
-  "khalq-afcal": true,
-  "qiraa-khalfa-imam": true,
-  "qurat-caynanyn": true,
-  "tarikh-kabir": true,
-  "tarikh-saghir": true,
-  "akhbar-muwaffaqiyyat": true,
-  "hadith-sufyan-ibn-cuyayna": true,
-  "jamharat-nasab-quraysh": true,
-  muntakhab: true,
-  "futuh-misr": true,
-  "ahwal-rijal": true,
-  "amarat-nubuwwa": true,
-  "musnad-bilal": true,
-  "kuna-wa-asma": true,
-  munfaridat: true,
-  "sahih-1": true,
-  tamyiz: true,
-  "juz-8": true,
-  "hadith-7": true,
-  "musnad-cumar-ibn-khattab": true,
-  mukhtasar: true,
-  "sharh-sunna": true,
-  "juz-sacdan": true,
-  "sirat-imam": true,
-  "tarikat-nabi": true,
-  "hadith-8": true,
-  "mahabba-li-llah": true,
-  "amali-wa-qiraa": true,
-  "juz-min-hadith": true,
-  "juz-ibn-cassam": true,
-  "nasikh-hadith": true,
-  "sunan-2": true,
-  "musnad-abi-umayya": true,
-  "musnad-cabd-allah-ibn-cumar": true,
-  "fitan-1": true,
-  "juz-hanbal": true,
-  "sunan-3": true,
-  "akhbar-shuyukh": true,
-  warac: true,
-  marasil: true,
-  "risala-ila-ahl-makka": true,
-  "sualat-ajurri": true,
-  "sunan-4": true,
-  "zuhd-6": true,
-  "hadith-10": true,
-  "diwan-abi-aswad-duali": true,
-  "ma-ruwiya-fi-hawd": true,
-  "adab-katib": true,
-  "anwa-fi-mawasim": true,
-  "ashriba-wa-ikhtilaf-fiha": true,
-  "cuyun-akhbar": true,
-  "gharib-hadith-1": true,
-  "gharib-quran-1": true,
-  "ikhtilaf-fi-lafz": true,
-  "islah-ghalat": true,
-  jarathim: true,
-  "macani-kabir": true,
-  macarif: true,
-  "masail-wa-ajwiba": true,
-  "risalat-khatt": true,
-  "shicr-wa-shucara": true,
-  "tawil-mukhtalaf-hadith": true,
-  "tawil-mushkil-quran": true,
-  "zuhd-7": true,
-  "macrifa-wa-tarikh": true,
-  "ansab-ashraf": true,
-  "futuh-buldan": true,
-  "mukhtasar-shamail-muhammadiyya": true,
-  "shamail-muhammadiyya": true,
-  "musnad-ibn-cawf": true,
-  "radd-cala-jahmiyya": true,
-  "masail-harb": true,
-  "ballaghat-nisa": true,
-  "fawaid-mucallala": true,
-  "tarikh-3": true,
-  ahwal: true,
-  "amr-bi-macruf": true,
-  awliya: true,
-  cuqubat: true,
-  cuzla: true,
-  "dhamm-dunya": true,
-  "dhamm-ghiba": true,
-  "dhamm-malahi": true,
-  "dhamm-muskir": true,
-  "fadail-ramadan": true,
-  "faraj-bacda-shidda": true,
-  "hamm-wa-huzn": true,
-  hilm: true,
-  "hilm-mucawiya": true,
-  "husn-zann-bi-llah": true,
-  "ikhlas-wa-niyya": true,
-  "istinac-macruf": true,
-  "kalam-layali": true,
-  "makaid-shaytan": true,
-  "makarim-akhlaq": true,
-  manamat: true,
-  "maqtal-cali": true,
-  "mudarat-nas": true,
-  "muhasabat-nafs": true,
-  muhtadarin: true,
-  "qada-hawaij": true,
-  "qasr-amal": true,
-  "qira-dayf": true,
-  "rasail-fi-tawakkul": true,
-  rida: true,
-  "riqqa-wa-buka": true,
-  sabr: true,
-  "sifat-janna": true,
-  "sifat-nar": true,
-  wajal: true,
-  "warac-1": true,
-  "zuhd-8": true,
-  "akhbar-tiwal": true,
-  "bughya-bahith": true,
-  "cawali-1": true,
-  "ahkam-quran-1": true,
-  "fadl-salat": true,
-  "amali-baghandi": true,
-  "tafsir-7": true,
-  taqfiya: true,
-  hamasa: true,
-  "hadith-11": true,
-  fadil: true,
-  "kamil-fi-lugha": true,
-  "nasab-cadnan": true,
-  "tacazi-1": true,
-  "gharib-hadith-2": true,
-  "ikram-dayf": true,
-  "risala-fi-quran-ghayr-makhluq": true,
-  bidac: true,
-  "ahad-wa-mathani": true,
-  awail: true,
-  diyat: true,
-  "jihad-1": true,
-  mudhakkir: true,
-  sunna: true,
-  "zuhd-9": true,
-  fakhir: true,
-  fasih: true,
-  "majalis-thaclab": true,
-  "qawacid-shicr": true,
-  "tarikh-wasit": true,
-  "jumca-wa-fadluha": true,
-  "musnad-abi-bakr": true,
-  "fadail-quran-1": true,
-  "ikhtilaf-culama": true,
-  "sunna-2": true,
-  "taczim-qadr-salat": true,
-  "fawaid-muntaqa": true,
-  badic: true,
-  "tabaqat-shucara": true,
-  "man-ismuh-camr": true,
-  waraqa: true,
-  carsh: true,
-  zahra: true,
-  tacliqat: true,
-  "ahkam-cidayn": true,
-  "dalail-nubuwwa": true,
-  "fadail-quran-2": true,
-  "fawaid-2": true,
-  "qadar-1": true,
-  "sifat-nifaq": true,
-  siyam: true,
-  "tabaqat-asma": true,
-  "tarikh-wa-asma-muhaddithin": true,
-  "juz-12": true,
-  "dalail-fi-gharib-hadith": true,
-  arbacuna: true,
-  "camal-yawm-wa-layla": true,
-  "dhikr-mudallisin": true,
-  "fadail-quran-3": true,
-  "juz-fi-masail-ibn-hanbal": true,
-  "juz-thalatha-wa-thalathina": true,
-  "mucjam-sahaba": true,
-  "musnad-cuthman": true,
-  "musnad-usama": true,
-  "cilal-ahadith": true,
-  "hadith-13": true,
-  zij: true,
-  "ahadith-2": true,
-  "juz-bi-riwayat-antaki": true,
-  "muntaqa-min-tabaqat": true,
-  "amali-3": true,
-  "musnad-ibn-abi-awfa": true,
-  jumca: true,
-  "khasais-amir-muminin-cali": true,
-  "majlisan-min-imla": true,
-  "sunan-sughra": true,
-  "tabaqat-1": true,
-  "tasmiyat-fuqaha-amsar": true,
-  "tasmiyat-man-lam-yarwi": true,
-  "amali-1": true,
-  "fawaid-3": true,
-  "akhbar-qudat": true,
-  "hadith-ibn-bashshar": true,
-  mafarid: true,
-  mucjam: true,
-  "musnad-12": true,
-  "musnad-13": true,
-  muntaqa: true,
-  "fadail-madina": true,
-  "fadail-makka-1": true,
-  "dhamm-thuqala": true,
-  "fadl-kilab": true,
-  murua: true,
-  dhariyya: true,
-  "kuna-wa-asma-1": true,
-  "amali-2": true,
-  "amr-bi-macruf-1": true,
-  "hathth-cala-tijara": true,
-  "qiraa-cinda-qubur": true,
-  "wuquf-wa-tarajjul": true,
-  "juz-13": true,
-  "fawaid-5": true,
-  "sahih-2": true,
-  "tawhid-1": true,
-  "macani-quran-2": true,
-  "tafsir-asma-allah": true,
-  "musnad-cumar": true,
-  "musnad-14": true,
-  "musnad-15": true,
-  ikhtiyarayn: true,
-  "akhbar-wa-hikayat": true,
-  mustakhraj: true,
-  bacth: true,
-  "musnad-caisha": true,
-  "usul-fi-nahw": true,
-  "hadith-muscab": true,
-  "hikayat-shucba": true,
-  "qubul-akhbar": true,
-  "awsat-fi-sunan": true,
-  ijmac: true,
-  iqnac: true,
-  "ishraf-cala-madhahib": true,
-  "acda-wa-nafs": true,
-  "adab-nafs": true,
-  "amthal-2": true,
-  "caql-wa-hawa": true,
-  manhiyyat: true,
-  "nawadir-usul": true,
-  "riyadat-nafs": true,
-  "mahasin-1": true,
-  "fawaid-wa-akhbar": true,
-  ishtiqaq: true,
-  "jamharat-lugha": true,
-  malahin: true,
-  "matar-wa-sahab": true,
-  mujtana: true,
-  "tacliq-min-amali": true,
-  "ahkam-quran-2": true,
-  "hadith-abi-jacfar": true,
-  "matn-caqida": true,
-  "sharh-macani-athar": true,
-  "sharh-mushkil-athar": true,
-  "taswiya-bayna-haddathana-wa-akhbarana": true,
-  "fawaid-6": true,
-  "ducafa-kabir": true,
-  tashbihat: true,
-  "ciyar-shicr": true,
-  "juz-15": true,
-  ibana: true,
-  "maqalat-islamiyyin": true,
-  "risala-ila-ahl-thughr": true,
-  majalis: true,
-  "sabca-fi-qiraat": true,
-  "amali-4": true,
-  muwashsha: true,
-  "akhbar-fakhkh": true,
-  "dhamm-muwaswisin": true,
-  "adab-nafica": true,
-  "adab-shafici": true,
-  "bayan-khata-bukhari": true,
-  "cilal-hadith": true,
-  "jarh-wa-tacdil": true,
-  "marasil-1": true,
-  "tafsir-11": true,
-  addad: true,
-  "idah-waqf-wa-ibtida": true,
-  "majlis-min-amali": true,
-  "mudhakkar-wa-muannath-1": true,
-  "sharh-qasaid-sabc": true,
-  "zahir-fi-macani": true,
-  "ciqd-farid": true,
-  "sharh-sunna-1": true,
-  "muntaqa-1": true,
-  "alfaz-1": true,
-  "gharib-quran-2": true,
-  "amali-6": true,
-  "duca-1": true,
-  "amali-7": true,
-  "fawaid-9": true,
-  "juz-19": true,
-  "ma-rawa-akabir": true,
-  "muntaqa-hadith": true,
-  intisar: true,
-  "maqsur-wa-mamdud": true,
-  mihan: true,
-  "tabaqat-culama": true,
-  "tafsir-13": true,
-  "tawhid-2": true,
-  iklil: true,
-  "sifat-jazirat-carab": true,
-  "mukhtasar-min-masail-ibn-hanbal": true,
-  "tarikh-raqqa": true,
-  "musnad-16": true,
-  "adab-qadi": true,
-  "fawaid-hadith-abi-cumayr": true,
-  "adab-kuttab": true,
-  "akhbar-abi-tammam": true,
-  "akhbar-radi": true,
-  "ashcar-awlad-khulafa": true,
-  "juz-min-ahadith": true,
-  "juz-21": true,
-  "akhbar-1": true,
-  "amali-8": true,
-  "huruf-macani": true,
-  "idah-cilal-nahw": true,
-  "ishtiqaq-asma-allah": true,
-  lamat: true,
-  "majalis-culama": true,
-  "naqd-shicr": true,
-  "cumdat-kuttab": true,
-  "icrab-quran": true,
-  "macani-quran-3": true,
-  "nasikh-wa-mansukh-4": true,
-  "qatc-wa-itinaf": true,
-  "siyasa-1": true,
-  "mucjam-1": true,
-  "qubl-wa-mucanaqa": true,
-  "zuhd-wa-sifat-zahidin": true,
-  "hadith-21": true,
-  "amali-9": true,
-  "usul-1": true,
-  "casharat-fi-gharib-lugha": true,
-  "maqsur-wa-mamdud-1": true,
-  "yaqutat-sirat": true,
-  majmuc: true,
-  "nuskha-5": true,
-  "tanbih-wa-ishraf": true,
-  "tashih-fasih": true,
-  "tarikh-6": true,
-  "fawaid-11": true,
-  "amali-10": true,
-  "dhikr-man-lahu-ayat": true,
-  "hadith-25": true,
-  "musnad-cumar-ibn-khattab-1": true,
-  "radd-cala-man-yaqul-quran-makhluq": true,
-  "akhbar-nahwiyyin": true,
-  "nawadir-fuqaha": true,
-  ittibac: true,
-  "muntaqa-min-musnad-muqillin": true,
-  "mucjam-sahaba-1": true,
-  "fawaid-14": true,
-  majruhin: true,
-  "mashahir-culama-amsar": true,
-  "rawdat-cuqala": true,
-  "sira-1": true,
-  thiqat: true,
-  "baric-fi-lugha": true,
-  "ittibac-1": true,
-  "maqsur-wa-mamdud-2": true,
-  "nuskhat-nabit": true,
-  "hadith-31": true,
-  "juz-bitaqa": true,
-  "fawaid-18": true,
-  "hadith-32": true,
-  "fawaid-19": true,
-  "juz-26": true,
-  "fadl-qiyam-layl": true,
-  tanbih: true,
-  "amthal-hadith": true,
-  "muhaddith-fasil": true,
-  "ahadith-tiwal": true,
-  "awail-2": true,
-  "duca-2": true,
-  "fadail-ramy": true,
-  "fadl-cashar-dhi-hijja": true,
-  "hadith-dabb": true,
-  "juz-hadith-ahl-basra": true,
-  "makarim-akhlaq-2": true,
-  "man-ismuh-cata": true,
-  "mucjam-awsat": true,
-  "mucjam-kabir": true,
-  "mucjam-saghir": true,
-  "musnad-shamiyyin": true,
-  "turuq-hadith-man-kadhdhaba-calayya": true,
-  "ziyadat-fi-jud-wa-sakha": true,
-  "mudhakkar-wa-muannath-2": true,
-  "muhibb-wa-mahbub": true,
-  "manaqib-shafici": true,
-  "camal-yawm-wa-layl": true,
-  "asami-man-rawa-canhum": true,
-  "kamil-fi-ducafa": true,
-  "man-wafaqat-kunyatuhu": true,
-  "juz-28": true,
-  afcal: true,
-  "akhbar-nahwiyyin-1": true,
-  "fawait-kitab-sibawayhi": true,
-  "sharh-kitab-sibawayhi": true,
-  "akhlaq-nabi": true,
-  "cawali-2": true,
-  "fawaid-23": true,
-  hikayat: true,
-  "tabaqat-muhaddithin": true,
-  tawbikh: true,
-  "fawaid-24": true,
-  "macani-qiraat": true,
-  "tahdhib-lugha": true,
-  "zahir-fi-gharib-alfaz-shafici": true,
-  "mutalif-wa-mukhtalif": true,
-  muwazana: true,
-  "hujja-fi-qiraat-sabca": true,
-  "icrab-thalathin-sura": true,
-  "laysa-fi-kalam-carab": true,
-  "ahkam-quran-3": true,
-  "fusul-fi-usul": true,
-  "sharh-mukhtasar-tahawi": true,
-  "ictiqad-aimmat-hadith": true,
-  "mucjam-2": true,
-  "tahdhib-mudawwana": true,
-  "bahr-culum": true,
-  "cuyun-masail-2": true,
-  "tanbih-ghafilin": true,
-  "ahadith-muntaqat": true,
-  "asma-man-yucraf-bi-kunya": true,
-  makhzun: true,
-  "man-wafaqa-ismuhu-ism-abihi": true,
-  tanbihat: true,
-  "juz-32": true,
-  "tanbih-wa-radd": true,
-  "hujja-li-qurra-sabca": true,
-  "idah-cadudi": true,
-  "masail-basriyyat": true,
-  "masail-caskariyyat": true,
-  "masail-halabiyyat": true,
-  "sharh-abyat": true,
-  "tacliqa-cala-kitab-sibawayhi": true,
-  "anwar-wa-mahasin-ashcar": true,
-  "asami-wa-kuna-1": true,
-  "cawali-malik-1": true,
-  "shicar-ashab-hadith": true,
-  tafric: true,
-  tawassut: true,
-  "qasida-nuniyya": true,
-  "gharaib-hadith-malik": true,
-  "hadith-shucba-1": true,
-  "tabaqat-nahwiyyin-wa-lughawiyyin": true,
-  "macani-akhbar": true,
-  "tacarruf-li-madhhab-ahl-tasawwuf": true,
-  talkhis: true,
-  "hadith-42": true,
-  "mabsut-fi-qiraat-cashar": true,
-  "arbacuna-1": true,
-  "fawaid-29": true,
-  "juz-nafic": true,
-  "min-hadith-1": true,
-  "mucjam-3": true,
-  "muntakhab-min-gharaib-malik": true,
-  rukhsa: true,
-  "cilal-nahw": true,
-  "amali-12": true,
-  "akhbar-musahhifin": true,
-  "masun-fi-adab": true,
-  "tashifat-muhaddithin": true,
-  "amthal-3": true,
-  "mufid-culum": true,
-  "ashcar-nisa": true,
-  "mucjam-shucara": true,
-  muwashshah: true,
-  afrad: true,
-  "ahadith-allati-khulifa-fiha-malik-ibn-anas": true,
-  "ahadith-muwatta": true,
-  "cilal-warida": true,
-  "dhikr-asma-tabicin": true,
-  "ducafa-2": true,
-  "fadail-sahaba-2": true,
-  "fawaid-afrad": true,
-  "ilzamat-wa-tatabbuc": true,
-  "mutalif-wa-mukhtalif-1": true,
-  ruyat: true,
-  "sualat-naysaburi": true,
-  "furqan-bayna-awliya": true,
-  "hamawiyya-kubra": true,
-  "hasana-wa-sayyia": true,
-  "hijab-mara": true,
-  hisba: true,
-  "huquq-al-bayt": true,
-  "iklil-1": true,
-  "iman-4": true,
-  "intisar-2": true,
-  "iqtida-sirat": true,
-  "istiqama-1": true,
-  "jamic-masail": true,
-  "jamic-rasail": true,
-  "jawab-fi-hilf": true,
-  "jawab-sahih": true,
-  "kalim-tayyib": true,
-  "sunan-5": true,
-  "tacliqat-cala-majruhin": true,
-  "sharh-abyat-sibawayhi": true,
-  "hadith-abi-qasim-cafiya": true,
-  "dhikr-man-ikhtalafa-culama": true,
-  "fadail-fatima": true,
-  "fadail-ramadan-1": true,
-  "juz-hadith-2": true,
-  "nasikh-hadith-wa-mansukhuh": true,
-  "sharh-madhahib": true,
-  "targhib-fi-fadail-acmal": true,
-  "tarikh-asma-ducafa": true,
-  "tarikh-asma-thiqat": true,
-  "amthal-saira": true,
-  "kashf-can-musawi-shicr-mutanabbi": true,
-  "muhit-fi-lugha": true,
-  ruznamaja: true,
-  "qut-qulub": true,
-  "dhabb-can-madhhab-imam-malik": true,
-  "jamic-fi-sunan": true,
-  "nawadir-wa-ziyadat": true,
-  "risala-2": true,
-  "fawaid-31": true,
-  "fawaid-muntaqa-1": true,
-  "hadith-44": true,
-  harbiyyat: true,
-  "mashyakha-4": true,
-  "lughat-fi-quran": true,
-  "ibana-kubra": true,
-  "ibtal-hiyal": true,
-  "sabcuna-fi-jihad": true,
-  "amali-13": true,
-  "mafatih-culum": true,
-  "aclam-hadith": true,
-  "bayan-icjaz-quran": true,
-  "cuzla-1": true,
-  "gharib-hadith-3": true,
-  "islah-ghalat-muhaddithin": true,
-  "macalim-sunan": true,
-  "shan-duca": true,
-  "fadail-tasmiyya": true,
-  "hilyat-muhadara": true,
-  "risala-wadiha": true,
-  "diyarat-1": true,
-  "fawaid-33": true,
-  "jalis-salih": true,
-  "ahadith-mia": true,
-  "alfaz-mahmuza": true,
-  carud: true,
-  "cilal-tathniya": true,
-  "cuqud-hamz": true,
-  khasais: true,
-  "lumac-fi-carabiyya": true,
-  "mubhij-fi-tafsir-asma-shucara": true,
-  "muhtasab-fi-tabyin": true,
-  munsif: true,
-  "sirr-sinacat-icrab": true,
-  "tamam-fi-tafsir-ashcar-hudhayl": true,
-  "sihah-taj-lugha": true,
-  "munsif-li-sariq": true,
-  "muntaqa-hadith-1": true,
-  "awail-3": true,
-  "diwan-macani": true,
-  "furuq-lughawiyya": true,
-  sinacatayn: true,
-  "wujuh-wa-nazair": true,
-  "hilyat-fuqaha": true,
-  "ittibac-2": true,
-  "makhadh-cilm": true,
-  "mucjam-maqayis": true,
-  "mujmal-lugha": true,
-  "mutakhayyar-alfaz": true,
-  sahibi: true,
-  "amali-15": true,
-  asami: true,
-  "fath-bab": true,
-  "fawaid-34": true,
-  "iman-3": true,
-  "macrifat-sahaba": true,
-  "musnad-ibn-adham": true,
-  "radd-cala-jahmiyya-1": true,
-  "tawhid-4": true,
-  "hadith-49": true,
-  "ahadith-4": true,
-  "cuyun-adilla": true,
-  maqamat: true,
-  "hadith-50": true,
-  "muntakhab-ahkam": true,
-  "qudwat-ghazi": true,
-  "tafsir-16": true,
-  gharibayn: true,
-  "mucjam-shuyukh": true,
-  "icjaz-quran": true,
-  "intisar-quran": true,
-  "tamhid-awail": true,
-  "taqrib-wa-irshad": true,
-  "minhaj-fi-shucab-iman": true,
-  "tarikh-culama-andalus": true,
-  "tahdhib-ajwiba": true,
-  "hujjat-qiraat": true,
-  "juz-36": true,
-  "mucjam-fi-mushtabah": true,
-  mumtic: true,
-  "macrifat-culum-hadith": true,
-  "madkhal-ila-iklil": true,
-  "madkhal-ila-sahih": true,
-  mustadrak: true,
-  "sualat-1": true,
-  "tasmiyat-man-akhrajahum": true,
-  "mushkil-hadith": true,
-  "tafsir-17": true,
-  "cuqala-majanin-1": true,
-  awham: true,
-  ghawamis: true,
-  mutawarin: true,
-  rubaci: true,
-  "fawaid-36": true,
-  "nasikh-wa-mansukh-5": true,
-  "adab-suhba": true,
-  "arbacuna-fi-tasawwuf": true,
-  "cuyub-nafs": true,
-  "dhikr-niswa-mutacabbidat": true,
-  "majlis-4": true,
-  "sualat-li-daruqutni": true,
-  "wasiyya-1": true,
-  "tafsir-muwatta": true,
-  "islam-zayd-ibn-haritha": true,
-  "musnad-muqallin": true,
-  "fawaid-ciraqiyyin": true,
-  "funun-cajaib": true,
-  "majalis-4": true,
-  "fawaid-39": true,
-  "amali-23": true,
-  "fawaid-40": true,
-  "lubab-fi-fiqh-shafici": true,
-  "tathbit-dalail-nubuwwa": true,
-  "tarikh-culama-ahl-misr": true,
-  "karamat-awliya": true,
-  "sharh-usul-ictiqad": true,
-  "adab-khawass": true,
-  inas: true,
-  "siyasa-2": true,
-  "hadith-57": true,
-  "arbacuna-buldaniyya": true,
-  "arbacuna-fi-fadail-dhikr-rabb": true,
-  "fawaid-41": true,
-  "nathr-durr": true,
-  "amali-marzuqi": true,
-  "azmina-wa-amkina": true,
-  "sharh-diwan-hamasa": true,
-  "hawamil-wa-shawamil": true,
-  "tahdhib-akhlaq": true,
-  tajarib: true,
-  "cuyun-masail-4": true,
-  ishraf: true,
-  macuna: true,
-  "sharh-risala": true,
-  "talqin-fi-fiqh": true,
-  "ahadith-5": true,
-  "qutb-surur": true,
-  "tawabic-wa-zawabic": true,
-  "kashf-wa-bayan": true,
-  "sualat-2": true,
-  "tarikh-jurjan": true,
-  "irshad-ila-sabil-rashad": true,
-  "mukhtasar-4": true,
-  "tajrid-1": true,
-  "juz-42": true,
-  "rijal-sahih-muslim": true,
-  "risalat-cukbari-fi-usul-fiqh": true,
-  "qanun-fi-tibb": true,
-  "hadith-60": true,
-  "ibana-1": true,
-  "mushkil-icrab-quran": true,
-  "jamc-wa-farq": true,
-  "mawqif-imam": true,
-  "risala-fi-ithbat": true,
-  "akhbar-thuqala": true,
-  "amali-31": true,
-  "dhikr-man-lam-yakun-lahu-illa-hadith-wahid": true,
-  "fadail-shahr-rajab": true,
-  "fadail-surat-ikhlas": true,
-  "karamat-awliya-1": true,
-  "siyasa-3": true,
-  "ahsan-ma-samictu": true,
-  "durar-hikam": true,
-  "fiqh-lugha": true,
-  "icjaz-wa-ijaz": true,
-  "khass-khass": true,
-  lutf: true,
-  "man-ghaba-canhu-mutrib": true,
-  muntahal: true,
-  "mutanabbi-wa-ma-lahu": true,
-  shakwa: true,
-  "sihr-balagha": true,
-  "tahsin-qabih": true,
-  tamthil: true,
-  "thimar-qulub": true,
-  "yatimat-dahr": true,
-  "fadail-ramy-1": true,
-  "farq-bayna-firaq": true,
-  burhan: true,
-  "ahadith-musnada": true,
-  "arbacuna-cala-madhhab-mutahaqqiqin": true,
-  "arbacuna-min-tibb": true,
-  "dalail-nubuwwa-1": true,
-  "ducafa-3": true,
-  "fadail-khulafa-rashidin": true,
-  "fadilat-cadilin": true,
-  "hilyat-awliya": true,
-  "imama-wa-radd-cala-rafida": true,
-  "juz-ahadith-can-abi-cali-sawwaf": true,
-  "macrifat-sahaba-1": true,
-  "majlis-min-amali-1": true,
-  "masanid-abi-yahya-kufi": true,
-  "muntakhab-min-hadith-yunus-ibn-cubayd": true,
-  "muntakhab-min-shucara": true,
-  "musnad-abi-hanifa-1": true,
-  "musnad-mustakhraj-cala-sahih-muslim": true,
-  "riyadat-abdan": true,
-  "sifat-janna-1": true,
-  "sifat-nifaq-1": true,
-  "tasmiyat-ruwat-can-fadl-ibn-dakin": true,
-  "tasmiyat-ruwat-can-sacid-ibn-mansur": true,
-  "tibb-nabawi": true,
-  taqwim: true,
-  "asma-khayl-carab": true,
-  "farhat-adib": true,
-  "amali-28": true,
-  "fawaid-45": true,
-  "hamasat-zurafa": true,
-  "fadail-quran-5": true,
-  "ibana-can-sariqat": true,
-  "fawaid-46": true,
-  masmucat: true,
-  "mukhtasar-nasih": true,
-  "muctamad-fi-usul-fiqh": true,
-  "amali-30": true,
-  furuq: true,
-  "ghayat-maqsud": true,
-  "ifham-yahud": true,
-  jamahir: true,
-  "tahqiq-ma-lil-hind": true,
-  "badic-1": true,
-  "hadith-61": true,
-  "fawaid-cawali": true,
-  "fawaid-muntaqat-1": true,
-  "sharh-shicr-mutanabbi": true,
-  takhrij: true,
-  "sharh-tasrif": true,
-  "tarikh-culama-nahwiyyin": true,
-  "hadith-malik": true,
-  "ahadith-7": true,
-  "ahruf-sabca": true,
-  "bayan-fi-cadd-ayy-quran": true,
-  "cilm-hadith": true,
-  "farq-bayna-dad-wa-za": true,
-  "jamic-bayan-fi-qiraat-sabc": true,
-  "muhkam-fi-naqt-masahif": true,
-  "muktafi-fi-waqf": true,
-  "muqnic-fi-rasm-masahif": true,
-  naqt: true,
-  "risala-wafiya": true,
-  "sunan-warida": true,
-  "tahdid-fi-itqan": true,
-  "taysir-fi-qiraat-sabc": true,
-  "dhikr-shuyukh": true,
-  "fadail-sham": true,
-  "risala-fi-radd": true,
-  "fawaid-47": true,
-  "irshad-fi-macrifa-culama": true,
-  "cawali-malik-2": true,
-  "rusum-dar-khilafa": true,
-  "tuhfat-umara": true,
-  "fusul-wa-ghayat": true,
-  "mucjiz-ahmad": true,
-  "risalat-ghufran": true,
-  "risalat-malaika": true,
-  "risalat-sahil-wa-shajih": true,
-  "risala-calawiyya": true,
-  "sharh-sahih-bukhari": true,
-  "furuq-fiqhiyya": true,
-  "aclam-nubuwwa-1": true,
-  "adab-dunya-wa-din": true,
-  "ahkam-sultaniyya": true,
-  "amthal-wa-hikam": true,
-  "durar-suluk": true,
-  "hawi-kabir": true,
-  "iqnac-fi-fiqh": true,
-  "nukat-wa-cuyun": true,
-  "tashil-nazar": true,
-  "sharh-mudawwana": true,
-  "fawaid-48": true,
-  "fadail-abi-bakr": true,
-  "jamic-li-masail-mudawwana": true,
-  "amali-33": true,
-  "majlis-fi-tawaduc": true,
-  "cumdat-kuttab-1": true,
-  "musnad-shihab": true,
-  "fath-cala-abi-fath": true,
-  "cunwan-fi-qiraat-sabca": true,
-  "akhlaq-wa-siyar": true,
-  "fisal-fi-milal": true,
-  "hajjat-widac": true,
-  "ihkam-fi-usul-ahkam": true,
-  "jamharat-ansab-1": true,
-  "jawamic-sira": true,
-  "maratib-ijmac": true,
-  "muhalla-bi-athar": true,
-  "nasikh-wa-mansukh-6": true,
-  "nubdha-kafiya": true,
-  "rasail-4": true,
-  "risala-fi-maratib-culum": true,
-  "taqrib-li-hadd-mantiq": true,
-  "tawq-hamama": true,
-  "diwan-102": true,
-  "cadad-fi-lugha": true,
-  "muhkam-wa-muhit": true,
-  mukhassas: true,
-  "sharh-mushkil": true,
-  "tarikh-8": true,
-  "diwan-103": true,
-  "masail-intiqad": true,
-  "majlis-7": true,
-  "nutaf-fifatawa": true,
-  "tacliqa-1": true,
-  "adab-mujalasa": true,
-  "bahjat-majalis": true,
-  durar: true,
-  "ikhtilaf-malik-wa-ashabihi": true,
-  "inbah-cala-qabail": true,
-  "insaf-1": true,
-  "intiqa-fi-fadail": true,
-  "isticab-fi-macrifat-ashab": true,
-  "istidhkar-jamic": true,
-  istighna: true,
-  "jamic-bayan-1": true,
-  "kafi-fi-fiqh": true,
-  "masail-mustaghriba": true,
-  "tamhid-muwatta": true,
-  "taqassi-hadith-fi-muwatta": true,
-  "cumda-fi-mahasin-shicr": true,
-  "asma-mubhama": true,
-  "bukhala-1": true,
-  "cawali-malik-3": true,
-  "dhikr-jahr-bi-basmala": true,
-  "faqih-wa-mutafaqqih": true,
-  "fasl-li-wasl-mudraj-fi-naql": true,
-  "ghunyat-multamis": true,
-  "hadith-sitta-min-tabicin": true,
-  "ihtijaj-bi-shafici": true,
-  "iqtida-cilm-camal": true,
-  "jamic-li-akhlaq-rawi-wa-adab-samic": true,
-  "juz-turuq-hadith-ibn-cumar-fi-tarai-hilal": true,
-  "kalam-cala-sifat": true,
-  "kifaya-fi-cilm-riwaya": true,
-  "muntakhab-min-zuhd-wa-raqaiq": true,
-  "muttafiq-wa-muftariq": true,
-  "muwaddih-awham": true,
-  "nasihat-ahl-hadith": true,
-  "qawl-fi-cilm-nujum": true,
-  "rihla-fi-talab-cilm": true,
-  "sabiq-wa-lahiq": true,
-  "sharaf-ashab-hadith": true,
-  "tali-talkhis-mutashabih": true,
-  "talkhis-mutashabih": true,
-  "taqyid-cilm": true,
-  "kamil-fi-qiraat": true,
-  "tarikh-baghdad": true,
-  tatfil: true,
-  "qut-mughtadhi": true,
-  "rawd-aniq-fi-fadl-siddiq": true,
-  "rih-nisrin": true,
-  "sabab-wadc": true,
-  "sawn-mantiq": true,
-  shamarikh: true,
-  "sharh-shawahid-mughni": true,
-  "sharh-sudur": true,
-  "sharh-sunan-nasai": true,
-  "sifat-sahib-dhawq-salim": true,
-  "tabaqat-huffaz": true,
-  "tabaqat-mufassirin": true,
-  "tabarri-min-macarrat-macarri": true,
-  "tadhkirat-mutasi": true,
-  "tadrib-rawi": true,
-  "lataif-isharat": true,
-  "nahw-qulub": true,
-  "risala-qushayriyya": true,
-  "dhayl-tarikh-mawlid-culama": true,
-  "sirr-fasaha": true,
-  "sharh-hamasa": true,
-  "dumyat-qasr": true,
-  "asbab-nuzul": true,
-  "sharh-diwan-mutanabbi": true,
-  "tafsir-basit": true,
-  "wajiz-fi-tafsir": true,
-  "fawaid-53": true,
-  "fawaid-54": true,
-  "sharh-muqaddima": true,
-  "kifayat-mutahaffiz": true,
-  "mustakhraj-1": true,
-  "radd-cala-man-yaqul": true,
-  "asrar-balagha": true,
-  "dalail-icjaz": true,
-  "darj-durar": true,
-  "miftah-fi-sarf": true,
-  "risala-shafiya": true,
-  "fadl-tahlil": true,
-  "risala-mughniyya": true,
-  wakhshiyyat: true,
-  "tabsira-1": true,
-  "hudud-fi-usul": true,
-  "ishara-fi-usul-fiqh": true,
-  "muntaqa-sharh-muwatta": true,
-  nasiha: true,
-  "tacdil-wa-tajrih": true,
-  "tahdhib-mustamirr": true,
-  "lumac-fi-usul-fiqh": true,
-  "macuna-fi-jadal": true,
-  "muhadhdhab-fi-fiqh": true,
-  "tabaqat-fuqaha": true,
-  "tabsira-fi-usul-fiqh": true,
-  "tanbih-fi-fiqh-shafici": true,
-  "ashcar-shucara": true,
-  "mashyakha-7": true,
-  "juz-44": true,
-  "burhan-fi-usul-fiqh": true,
-  "ghiyath-umam": true,
-  "ijtihad-min-talkhis": true,
-  "lamc-adilla": true,
-  "nihayat-matlab": true,
-  "talkhis-fi-usul-fiqh": true,
-  waraqat: true,
-  "nukat-fi-quran": true,
-  "hafawat-nadira": true,
-  "arbacuna-fi-dalail-tawhid": true,
-  "dhamm-kalam": true,
-  "manazil-sairin": true,
-  "wafayat-misriyyin": true,
-  "juz-46": true,
-  "dahd-shubuhat": true,
-  "intisar-li-hizb-allah": true,
-  "radd-cala-burda": true,
-  "rasail-wa-fatawa-2": true,
-  "tasis-taqdis": true,
-  "tacliq-latif": true,
-  "bayan-kalimat-tawhid": true,
-  "bayan-mahabba": true,
-  "fath-majid": true,
-  "iman-5": true,
-  "kashf-ma-alqahu-iblis": true,
-  "maqamat-5": true,
-  "masail-wa-fatawa-najdiyya": true,
-  "matlab-hamid": true,
-  "mabsut-1": true,
-  "sharh-siyar-kabir": true,
-  "usul-2": true,
-  "amali-38": true,
-  "diwan-ahkam-kubra": true,
-  "hadiyyat-ahya": true,
-  "sharh-mucallaqat": true,
-  "ithar-insaf": true,
-  "tarikh-irbil": true,
-  "turath-fi-tafsir": true,
-  "tawjih-lumac": true,
-  "tarikh-dunaysir": true,
-  "muntakhab-siyaq": true,
-  "bughyat-nuqqad": true,
-  "hidayat-murtab": true,
-  "jamal-qurra": true,
-  "safar-sacada": true,
-  "ahadith-mukhatara": true,
-  "ahadith-sahiha-mimma-rawahu-muslim": true,
-  "bulghat-talib-hathith-1": true,
-  cudda: true,
-  "dhikr-musahafa": true,
-  "fadail-acmal": true,
-  "fadail-bayt-maqdis-1": true,
-  "hadith-ibn-abi-makarim": true,
-  "ikhtisas-quran": true,
-  "ittibac-sunan": true,
-  "juz-awham": true,
-  "khamsat-ahadith-musalsalat": true,
-  "manaqib-jacfar-ibn-abi-talib": true,
-  "min-hadith-ibn-yazid-muqri": true,
-  "muntaqa-hadith-cabdawi": true,
-  "muntaqa-min-masmucat-marw": true,
-  "muwafaqat-cawali": true,
-  "muwafaqat-fi-mashayikh-imam-ahmad": true,
-  "nahy-can-sabb-ashab": true,
-  "ruwat-arbacat-cashar": true,
-  "ruwat-can-muslim": true,
-  "sifat-janna-2": true,
-  "sifat-nabi-1": true,
-  "sunan-wa-ahkam-can-mustafa": true,
-  "turuq-hadith-nabi": true,
-  "ziyadat-cala-kabair": true,
-  "naqd-sahih": true,
-  "tahqiq-munif-rutba": true,
-  "tahqiq-murad": true,
-  "tanbihat-1": true,
-  "asila-wa-ajwiba-fi-icrab-quran": true,
-  "awdah-masalik": true,
-  "ictirad-shart": true,
-  "mabahith-murdiyya": true,
-  "masail-safariyya": true,
-  "matn-qatr-nada": true,
-  "sharh-hudud": true,
-  "badaic-silk": true,
-  "taj-wa-iklil-li-mukhtasar-khalil": true,
-  "rafc-niqab": true,
-  "cuqud-juman": true,
-  "cuqud-zabarjad": true,
-  cushariyyat: true,
-  "dafc-tashnic": true,
-  "dhamm-maks": true,
-  "dhamm-qada": true,
-  "durar-muntathira": true,
-  "durr-manthur": true,
-  "fadd-wica": true,
-  fanid: true,
-  "fath-kabir": true,
-  "ghurar-fi-fadail-cumar": true,
-  habaik: true,
-  "hamc-hawamic": true,
-  "hawi-li-fatawi": true,
-  "husn-muhadara": true,
-  "husn-samt": true,
-  "ifadat-khabar": true,
-  "iklil-2": true,
-  "iqtirah-1": true,
-  "iscaf-mubatta": true,
-  "itmam-diraya": true,
-  itqan: true,
-  "ittibac-4": true,
-  izdihar: true,
-  "idah-maknun": true,
-  "tafsir-30": true,
-  "nur-yaqin": true,
-  "mukhtasar-fath-rabb-arbab": true,
-  "jawahir-1": true,
-  "madkhal-ila-madhhab-ahmad": true,
-  munadama: true,
-  "badhl-majhud-2": true,
-  "fasl-maqal": true,
-  "mucjam-ma-istacjama": true,
-  "iclam-macalim": true,
-  "ikhbar-ahl-rusukh": true,
-  "jamic-masanid": true,
-  "tarikh-mukhtasar-duwal": true,
-  "ghusun-yanica": true,
-  "jughrafiya-1": true,
-  mughrib: true,
-  muqtataf: true,
-  "madarij-salikin": true,
-  "dhayl-thamarat-awraq": true,
-  "khizanat-adab": true,
-  sharh: true,
-  "fath-rahman-1": true,
-  "fath-mutacal": true,
-  "ajwiba-muhaqqaqa": true,
-  "fawaid-cajiba": true,
-  "radd-muhtar": true,
-  "simt-laali-fi-sharh-amali": true,
-  "tanbih-cala-awham": true,
-  "amali-39": true,
-  "akhbar-wa-ashcar": true,
-  "jadhwat-muqtabis": true,
-  "jamc-bayna-sahihayn": true,
-  tadhkira: true,
-  "tafsir-gharib": true,
-  "masmucat-1": true,
-  "fawaid-cawali-2": true,
-  "ictiqad-qadiri": true,
-  "intisar-ashab-hadith": true,
-  "qawatic-adilla": true,
-  "tafsir-quran-3": true,
-  "tacliq-cala-muwatta": true,
-  "amali-40": true,
-  "amali-41": true,
-  "juz-48": true,
-  "majlis-yawm-jumca": true,
-  "fawaid-58": true,
-  "hadith-sufyan-ibn-cuyayna-1": true,
-  khulciyyat: true,
-  "muntaqa-min-khulciyyat": true,
-  "mukhtasar-tabyin": true,
-  "aqdiyat-rasul": true,
-  "alqab-sahaba": true,
-  "taqyid-muhmal": true,
-  "juz-49": true,
-  "rawdat-qudat": true,
-  "masaric-cushshaq": true,
-  "tathqif-lisan": true,
-  "bahr-madhhab": true,
-  "diwan-hamasa": true,
-  "sharh-qasaid": true,
-  "dharica-fi-makarim-sharica": true,
-  mufradat: true,
-  "muhadarat-udaba": true,
-  "tafsil-nashatayn": true,
-  "tafsir-19": true,
-  "ahkam-quran-4": true,
-  "asnaf-maghrurin": true,
-  "bidayat-hidaya": true,
-  fadaih: true,
-  "ihya-culum-din": true,
-  "iqtisad-1": true,
-  "jawahir-quran": true,
-  "kimiya-sacada": true,
-  "macarij-quds": true,
-  "mahk-nazar": true,
-  "majmucat-rasail-1": true,
-  mankhul: true,
-  "maqsad-asna": true,
-  "micyar-cilm": true,
-  "mishkat-anwar": true,
-  "mizan-camal": true,
-  munqidh: true,
-  mustasfa: true,
-  "qawacid-caqaid": true,
-  "shifa-ghalil": true,
-  tahafut: true,
-  "tibr-masbuk": true,
-  wasit: true,
-  "fawaid-60": true,
-  "hilyat-culama": true,
-  "ansab-muttafiqa": true,
-  "atraf-gharaib": true,
-  "dhakhirat-huffaz": true,
-  "idah-ishkal": true,
-  "macrifat-tadhkira": true,
-  manthur: true,
-  "masalat-culuww-wa-nuzul": true,
-  "masalat-tasmiyya": true,
-  "tadhkirat-huffaz": true,
-  "fawaid-61": true,
-  firdaws: true,
-  "hidaya-2": true,
-  "tamhidh-fi-usul-fiqh": true,
-  "manhal-rawi": true,
-  "tadhkirat-samic": true,
-  "tahrir-ahkam-1": true,
-  "tanqih-munazara": true,
-  "nihayat-arab": true,
-  "mawrid-fi-camal-mawlid": true,
-  "riyad-afham": true,
-  "masail-fiqhiyya-1": true,
-  "mulhaq-fatawa": true,
-  "cuyun-athar": true,
-  "nafh-shadhi": true,
-  "madkhal-2": true,
-  "masail-tahlil-haid": true,
-  "nasikh-quran": true,
-  ihsan: true,
-  "tuhfat-sadiq": true,
-  "idah-fi-culum-balagha": true,
-  marasid: true,
-  "qawacid-usul": true,
-  "manzuma-tabriziyya": true,
-  "kanz-fi-qiraat-cashr": true,
-  "idah-dalail": true,
-  "qawanin-fiqhiyya": true,
-  "taqrib-wusul": true,
-  "tamhid-wa-bayan": true,
-  "mishkat-masabih": true,
-  "lubab-tawil": true,
-  "majid-fi-icrab": true,
-  "muntaqa-min-fawaid": true,
-  "tahdhib-kamal": true,
-  "tarjamat-maslama": true,
-  "tuhfat-ashraf": true,
-  "rawdat-taqrir": true,
-  "tabyin-haqaiq": true,
-  "futuh-ghayb": true,
-  khulasa: true,
-  "cuqud-durriyya": true,
-  "nafh-tib": true,
-  "sira-halabiyya": true,
-  "adab-culama": true,
-  "rawda-rayya": true,
-  "cumdat-talib-1": true,
-  "kashf-zunun": true,
-  "sullam-wusul": true,
-  "ithaf-dhawi-itqan": true,
-  "maraqi-falah-sharh-nur-idah": true,
-  "nur-idah": true,
-  "sacadat-majid": true,
-  "tajaddud-masarrat": true,
-  "hashiya-cala-tafsir-baydawi-1": true,
-  "rayhanat-alibba": true,
-  "sharh-durrat-ghawwas": true,
-  "rihlat-shita": true,
-  "durr-thamin-1": true,
-  "sharh-mayyara": true,
-  "subh-munabbi": true,
-  "majmac-anhur": true,
-  "manzuma-bayquniyya": true,
-  "akhsar-mukhtasarat": true,
-  "bughyat-mustafid": true,
-  "mukhtasar-ifadat": true,
-  "dhayl-lubb-lubab": true,
-  "hashiya-cala-muntaha-iradat": true,
-  "durr-mukhtar": true,
-  "muctiyat-aman": true,
-  shadharat: true,
-  "tanbih-khillan": true,
-  "khizanat-adab-1": true,
-  "sharh-abyat-mughni": true,
-  taysir: true,
-  "jamc-fawaid": true,
-  "silat-khalaf": true,
-  "tuhaf-rabbaniyya": true,
-  "murqisat-wa-mutribat": true,
-  hashiya: true,
-  "risalat-ayy-mushaddada": true,
-  "amali-46": true,
-  "juz-fihi-dhikr": true,
-  "macrifat-asami-ardaf": true,
-  "man-casha": true,
-  funun: true,
-  "fusul-adab": true,
-  "tadhkira-fi-fiqh": true,
-  "wadih-fi-usul-fiqh": true,
-  "juz-51": true,
-  rubaciyyat: true,
-  "afcal-2": true,
-  "durra-khatira": true,
-  "durrat-ghawwas": true,
-  "maqamat-1": true,
-  "mulhat-icrab": true,
-  "majmac-amthal": true,
-  "bayan-wa-tahsil": true,
-  "masail-4": true,
-  muqaddimat: true,
-  "radd-cala-man-dhahaba": true,
-  hawadith: true,
-  sila: true,
-  "inba-fi-tarikh-khulafa": true,
-  "siraj-muluk": true,
-  hadaiq: true,
-  "hulal-fi-sharh": true,
-  "insaf-fi-tanbih": true,
-  iqtidab: true,
-  "mushkilat-muwatta": true,
-  "rasail-fi-lugha": true,
-  "takmilat-tarikh-tabari": true,
-  "dhayl-dhayl-tarikh-mawlid-culama": true,
-  "hadith-71": true,
-  "ahadith-11": true,
-  "ictiqad-2": true,
-  "masail-5": true,
-  "tabaqat-hanabila": true,
-  "hadith-72": true,
-  "matmah-anfus": true,
-  ima: true,
-  "rasail-falsafiyya-1": true,
-  "ahadith-subaciyyat": true,
-  "cawali-malik-4": true,
-  "juz-tuhfat-cid-fitr": true,
-  "mashyakha-8": true,
-  "cawali-4": true,
-  "dalail-nubuwwa-4": true,
-  "hujja-fi-bayan-mahajja": true,
-  "icrab-quran-2": true,
-  musalsalat: true,
-  "siyar-salaf": true,
-  "targhib-wa-tarhib": true,
-  "idah-2": true,
-  "muclim-bi-fawaid-muslim": true,
-  "sharh-talqin": true,
-  "sharh-adab-qadi": true,
-  "asas-balagha": true,
-  "atwaq-dhahab": true,
-  "faiq-fi-gharib": true,
-  "jibal-wa-amkina-wa-miyah": true,
-  "kalim-nawabigh": true,
-  "maqamat-2": true,
-  "mufassal-fi-suncat-icrab": true,
-  "mustaqsa-fi-amthal": true,
-  qustas: true,
-  "rabic-abrar": true,
-  "ruus-masail": true,
-  "ma-jaa-cala-facaltu": true,
-  "sharh-adab-katib": true,
-  "takmila-cala-durrat-ghawass": true,
-  "amali-52": true,
-  "fawaid-63": true,
-  "mizan-usul": true,
-  "tuhfat-fuqaha": true,
-  "iqnac-fi-qiraat": true,
-  dhakhira: true,
-  "amali-53": true,
-  "ma-lam-yunshar-1": true,
-  "mukhtarat-shucara": true,
-  "ahkam-quran-5": true,
-  "cawasim-min-qawasim": true,
-  mahsul: true,
-  "masalik-fi-sharh-muwatta": true,
-  "nasikh-wa-mansukh-7": true,
-  "qabas-fi-sharh-muwatta": true,
-  "qanun-tawil": true,
-  "icrab-quran-3": true,
-  ghunya: true,
-  "ikmal-muclim": true,
-  "ilmac-ila-macrifa": true,
-  "mashariq-anwar": true,
-  shifa: true,
-  "tanbihat-mustanbita": true,
-  "tartib-madarik": true,
-  "arbacuna-3": true,
-  "adab-hisba": true,
-  badhl: true,
-  tariqa: true,
-  "bayan-fi-madhhab-shafici": true,
-  "intisar-fi-radd": true,
-  "ifsah-1": true,
-  "ikhtilaf-aimma": true,
-  "ghunya-1": true,
-  "tadhkira-hamduniyya": true,
-  "carus-ajza": true,
-  "fawaid-65": true,
-  "adab-imla": true,
-  "ansab-1": true,
-  "fadail-sham-1": true,
-  "muntakhab-3": true,
-  tahbir: true,
-  "mucjam-5": true,
-  "mujibat-janna": true,
-  "safat-mulah": true,
-  "lubab-ansab": true,
-  "tatimmat-siwan-hikma": true,
-  murtajal: true,
-  "matalic-anwar": true,
-  "nukat-casriyya": true,
-  "arbacuna-min-masanid": true,
-  "qurt-cala-kamil": true,
-  "hadith-73": true,
-  "hur-cin": true,
-  "shams-culum": true,
-  cumda: true,
-  "ahadith-can-jacfar-sarraj": true,
-  "ahadith-cidiyya": true,
-  "ahadith-wa-hikayat": true,
-  "arbacuna-buldaniyya-2": true,
-  "arbacuna-fi-haqq-fuqara": true,
-  "hadith-can-hakim-kufa": true,
-  "hadith-musahafa": true,
-  "juz-bi-intikhab": true,
-  "juz-thalathat-ahadith": true,
-  "majalis-khamsa": true,
-  "mashyakha-11": true,
-  "mashyakha-baghdadiyya": true,
-  "mucjam-safar": true,
-  "muntaqa-min-safina-baghdadiyya": true,
-  "muqaddimat-imla-istidhkar": true,
-  "qasida-1": true,
-  "shart-qiraa": true,
-  tuyuriyyat: true,
-  "wajiz-1": true,
-  "arbacun-calawiyya": true,
-  "asrar-carabiyya": true,
-  "bulgha-fi-farq": true,
-  "insaf-fi-masail": true,
-  "nuzhat-alibba": true,
-  "madkhal-1": true,
-  "sharh-fasih": true,
-  "athar-marwiyya": true,
-  "ghawamid-asma": true,
-  "ma-ruwiya-fi-hawd-1": true,
-  mustaghithin: true,
-  "amali-54": true,
-  "dhikr-ibn-abi-dunya": true,
-  "dhikr-ibn-mandah": true,
-  "kalam-cala-hadith-istilqa": true,
-  "khasais-musnad-ahmad": true,
-  lataif: true,
-  "majmuc-mughith": true,
-  "muntaha-raghabat": true,
-  "nuzhat-huffaz": true,
-  "rubaci-tabicin": true,
-  "faraid-1": true,
-  "nataij-fikr-fi-nahw": true,
-  "rawd-unuf": true,
-  "ahkam-sharciyya": true,
-  "ahkam-sughra": true,
-  "ahkam-wusta": true,
-  "caqiba-fi-dhikr-mawt": true,
-  "jamc-bayn-sahihayn": true,
-  "hayy-ibn-yaqzan": true,
-  "fi-tacrib": true,
-  "ghalat-ducafa-min-fuqaha": true,
-  "hawashi-cala-durrat-ghawass": true,
-  "masail-manthura-fi-tafsir": true,
-  "badic-fi-naqd-shicr": true,
-  "ictibar-1": true,
-  "lubab-adab-1": true,
-  "manazil-wa-diyar": true,
-  amakin: true,
-  "cujalat-mubtadi": true,
-  "ictibar-fi-nasikh-wa-mansukh": true,
-  "tabaqat-fuqaha-yaman": true,
-  subaciyyat: true,
-  "badaic-sanaic": true,
-  "hirz-amani": true,
-  "manhaj-masluk": true,
-  "nihayat-rutba": true,
-  "radd-cala-nuhat": true,
-  "taqwim-nazar": true,
-  bidaya: true,
-  "hidaya-fi-sharh-bidaya": true,
-  "usul-din-2": true,
-  "arbacuna-jilaniyya": true,
-  "bidayat-mujtahid": true,
-  "daruri-fi-usul-fiqh": true,
-  "fasl-maqal-1": true,
-  "risalat-nafs": true,
-  "talkhis-burhan": true,
-  "talkhis-cibara": true,
-  "talkhis-khitaba": true,
-  "talkhis-qiyas": true,
-  "talkhis-safsata": true,
-  adhkiya: true,
-  "akhbar-hamqa-wa-mughaffalin": true,
-  "akhbar-nisa": true,
-  "bahr-dumuc": true,
-  "birr-wa-sila-1": true,
-  "birr-walidayn": true,
-  "bustan-wacizin": true,
-  "cilal-mutanahiyya": true,
-  "dhamm-hawa": true,
-  "ducafa-wa-matrukin-1": true,
-  "fadail-quds": true,
-  "funun-afnan": true,
-  "gharib-hadith-4": true,
-  hathth: true,
-  "hifz-cumr": true,
-  "kashf-mushkil": true,
-  laali: true,
-  "lataif-1": true,
-  "manaqib-imam-ahmad": true,
-  "manthur-1": true,
-  "mashyakha-12": true,
-  mawducat: true,
-  mudhish: true,
-  mujtaba: true,
-  muntazam: true,
-  muqliq: true,
-  "musaffa-bi-akuff": true,
-  "musalsalat-1": true,
-  "muthir-gharam": true,
-  "nawasikh-quran": true,
-  "nuzhat-acyun": true,
-  "qussas-wa-mudhakkirin": true,
-  "sayd-khatir": true,
-  "sifat-safwa": true,
-  "tabsira-2": true,
-  "taczim-futya": true,
-  "tadhkira-fi-wacz": true,
-  "tadhkirat-arib": true,
-  "tahqiq-fi-ahadith-khilaf": true,
-  "talbis-iblis": true,
-  "talqih-fuhum": true,
-  "tanbih-naim": true,
-  "tanwir-ghabash": true,
-  "taqwim-lisan": true,
-  "tarikh-bayt-muqaddas": true,
-  "thibat-cinda-mamat": true,
-  "zad-masir": true,
-  "ziraf-wa-mutamajinin": true,
-  "muntaha-talab": true,
-  "hazz-ghalasim": true,
-  "ahadith-jamacili": true,
-  "akhbar-dajjal": true,
-  "akhbar-hasan-basri": true,
-  "akhbar-salat": true,
-  "amr-bi-macruf-3": true,
-  "caqida-2": true,
-  "cumdat-ahkam": true,
-  "fadail-cumar-ibn-khattab": true,
-  "fadail-shahr-ramadan": true,
-  "hadith-ifk": true,
-  "iqtisad-fi-ictiqad": true,
-  "juz-ahadith-shicr": true,
-  "misbah-fi-cuyun-sihah": true,
-  "nihayat-murad": true,
-  "tahrim-qatl": true,
-  "targhib-fi-duca": true,
-  "tawakkul-2": true,
-  "tawhid-li-llah": true,
-  "zawaj-abi-cas-bi-zaynab": true,
-  "jawhar-naqi": true,
-  "tariq-hijratayn": true,
-  "imla-mukhtasar": true,
-  "ictiqadat-1": true,
-  "macalim-usul-din": true,
-  "mafatih-ghayb": true,
-  "mahsul-1": true,
-  "badic-2": true,
-  "jamic-usul": true,
-  "nihaya-fi-gharib": true,
-  shafi: true,
-  "lataif-dhakhira": true,
-  "muqaddima-fi-nahw": true,
-  "ahadith-16": true,
-  "juz-58": true,
-  "tahrir-maqal": true,
-  "dar-tiraz": true,
-  "hamasa-maghribiyya": true,
-  "mughrib-fi-tartib-mucrib": true,
-  "ahadith-muqtabasa": true,
-  "arbacuna-6": true,
-  "juz-fihi-turuq-cabd-rahman": true,
-  "badaic-badaih": true,
-  "gharaib-tanbihat": true,
-  "ittifaq-mabani": true,
-  "midmar-haqaiq": true,
-  "rihla-2": true,
-  "murshid-zuwwar": true,
-  "muhit-burhani": true,
-  "tahqiq-wa-bayan": true,
-  "ciqd-jawahir": true,
-  "furuq-1": true,
-  "icrab-lamiyat-shanfara": true,
-  "icrab-ma-yushkil": true,
-  "ittihaf-hadith": true,
-  lubab: true,
-  "masail-khilafiyya": true,
-  "sharh-diwan-mutanabbi-1": true,
-  tabyin: true,
-  "tibyan-fi-icrab-quran": true,
-  "arbacina-fi-jihad": true,
-  "sharh-maqamat-hariri": true,
-  "arbacuna-fi-manaqib-ummahat": true,
-  "bulghat-talib-hathith": true,
-  "cumdat-fiqh": true,
-  "cumdat-hazim": true,
-  "khulasat-wafa": true,
-  "maqalat-musfira": true,
-  "wafa-wafa": true,
-  "amr-bi-ittibac": true,
-  "arbacuna-min-riwayat-malik-can-nafic": true,
-  "ashbah-1": true,
-  "asma-mudallisin": true,
-  "dhamm-tawil": true,
-  "fadl-yawm-tarwiyya": true,
-  "hikayat-munazara": true,
-  "ithbat-sifat-culuww": true,
-  "kafi-fi-fiqh-1": true,
-  "lumcat-ictiqad": true,
-  mughni: true,
-  "maakhidh-cala-shurrah": true,
-  "amali-56": true,
-  "jamic-ummahat": true,
-  "kafiya-fi-cilm-nahw": true,
-  "shafiya-fi-cilm-tasrif": true,
-  "jumal-fi-mantiq": true,
-  mucjib: true,
-  "manar-huda": true,
-  "mirat-zaman": true,
-  "fawaid-69": true,
-  "ikhtisar-sahih-bukhari": true,
-  mufhim: true,
-  "falak-dair": true,
-  "takhrij-furuc": true,
-  "hadith-mutabayicin-bi-khiyar": true,
-  "jawab-fi-jarh-wa-tacdil": true,
-  "juz-fi-cikrima": true,
-  "mukhtasar-sahih-muslim": true,
-  "mukhtasar-sunan-abi-dawud": true,
-  "takmila-li-wafayat-naqala": true,
-  "targhib-wa-tarhib-1": true,
-  "nagrat-ighrid": true,
-  "mudhakara-fi-alqab": true,
-  "hulla-siyara": true,
-  "ictab-kitab": true,
-  "takmila-li-sila": true,
-  "tuhfat-qadim": true,
-  "hamasa-1": true,
-  "mashyakha-16": true,
-  "bughyat-talab": true,
-  "zubdat-halab": true,
-  "fawaid-fi-ikhtisar-maqasid": true,
-  "ghaya-fi-ikhtisar-nihaya": true,
-  "imam-fi-bayan": true,
-  "maqasid-ricaya": true,
-  "munyat-sul": true,
-  "qawacid-ahkam": true,
-  "tafsir-quran-4": true,
-  muyassar: true,
-  "ghurar-fawaid": true,
-  "nuzhat-nazir-1": true,
-  "bacith-cala-inkar-bidac": true,
-  "ibraz-macani": true,
-  "khutbat-muammal": true,
-  "mukhtasar-muammal": true,
-  "murshid-wajiz": true,
-  rawdatayn: true,
-  "sharh-hadith-muqtafa": true,
-  "anmudhaj-jalil": true,
-  "mukhtar-sihah": true,
-  "tuhfat-muluk": true,
-  "intikhab-li-kashf-abyat": true,
-  "cuyun-anba": true,
-  takhjil: true,
-  "darair-shicr": true,
-  "mumtic-kabir-fi-tasrif": true,
-  "iclam-bima-fi-din-nasara": true,
-  "jamic-li-ahkam-quran": true,
-  "tadhkira-fi-ahwal-mawta": true,
-  "ahadith-sitta-ciraqiyya": true,
-  "juz-66": true,
-  "alfaz-mukhtalifa": true,
-  alfiyya: true,
-  "ictimad-fi-nazair": true,
-  "ijaz-tacrif": true,
-  "ikmal-aclam": true,
-  "macani-abniyat-asma": true,
-  "sharh-kafiya": true,
-  "sharh-tashil-fawaid": true,
-  "shawahid-tawdih": true,
-  "tashil-fawaid": true,
-  "zakhair-fi-lugha": true,
-  "nur-qabas": true,
-  "rawdat-mustabin": true,
-  "durr-thamin": true,
-  "adab-fatwa": true,
-  adhkar: true,
-  "arbacuna-nawawiyya": true,
-  "bustan-carifin-1": true,
-  "daqaiq-minhaj": true,
-  "ictiqad-salaf-fi-huruf": true,
-  "idah-fi-manasik-hajj": true,
-  "ijaz-fi-sharh-sunan-abi-dawud": true,
-  "irshad-tullab-haqaiq": true,
-  "khulasat-ahkam": true,
-  "majmuc-4": true,
-  "minhaj-fi-sharh-muslim": true,
-  "minhaj-talibin": true,
-  "rawdat-talibin": true,
-  "riyad-salihin": true,
-  "tahdhib-asma": true,
-  "tahrir-alfaz": true,
-  "matn-shudhur-dhahab": true,
-  "mughni-labib": true,
-  "sharh-qatr-nada": true,
-  "sharh-shudhur-dhahab": true,
-  "takhlis-shawahid": true,
-  "ghurra-munifa": true,
-  taqrib: true,
-  "tibyan-2": true,
-  "usul-wa-dawabit": true,
-  "sabr-1": true,
-  "kashf-asrar": true,
-  "talkhis-cibara-1": true,
-  "sharh-tacrif": true,
-  "wafayat-acyan": true,
-  "sharh-kabir-1": true,
-  tahsil: true,
-  "athar-bilad": true,
-  mutawari: true,
-  "aclaq-khatira": true,
-  "ciqd-manzum": true,
-  "dhakhira-1": true,
-  "furuq-2": true,
-  "ihkam-fi-tamyiz-fatawa": true,
-  "nafais-usul": true,
-  "sharh-tanqih": true,
-  umniyya: true,
-  "nashwat-tarab": true,
-  "anwar-tanzil": true,
-  "tuhfat-abrar": true,
-  "ahadith-safar": true,
-  "ithaf-zair": true,
-  "sharh-alfiyya": true,
-  "lubab-fi-jamc": true,
-  "sharh-shafiya-ibn-hajib": true,
-  "shamil-fi-sinaca-tibbiyya": true,
-  "sharh-tashrih-qanun": true,
-  "mukhtasar-minhaj-qasidin": true,
-  "tarikh-mustabsir": true,
-  "fihrist-4": true,
-  "tuhfat-majd": true,
-  "fadail-kitab-jamic": true,
-  "aqsa-amal": true,
-  dhakhair: true,
-  "khulasat-siyar-1": true,
-  "riyad-nadira": true,
-  "mumtic-fi-sharh-muqnic": true,
-  "sifat-fatwa": true,
-  "mashyakha-19": true,
-  "qawacid-tafsir-ahlam": true,
-  "mufarrij-kurub": true,
-  "matn-muwattat-fasih": true,
-  gharamiyya: true,
-  "mukhtasar-khilafiyyat": true,
-  "maqamat-3": true,
-  "mashyakha-20": true,
-  "arbacuna-hadithan-tusaciyat-isnad": true,
-  "ihkam-ahkam-sharh-cumdat-ahkam": true,
-  "ilmam-bi-ahadith": true,
-  iqtirah: true,
-  "sharh-arbacina-nawawiyya": true,
-  "sharh-ilmam-bi-ahadith": true,
-  "dhayl-wa-takmila": true,
-  "durr-nathir": true,
-  "ahadith-cawali": true,
-  "arbacuna-abdal": true,
-  "juz-musafahat-muslim": true,
-  "mucjam-shuyukh-2": true,
-  muwafaqiyyat: true,
-  tasalli: true,
-  "burhan-2": true,
-  "malak-tawil-qatic": true,
-  "matlac-cala-abwab-muqnic": true,
-  "fakhri-fi-adab": true,
-  "kanz-daqaiq": true,
-  "tafsir-23": true,
-  "durr-farid": true,
-  "nafahat-nasamat": true,
-  kifaya: true,
-  "kafi-2": true,
-  "miftah-tariq": true,
-  "miftah-tariq-ila-suluk-tahqiq": true,
-  "miftah-tariq-muhibbin": true,
-  "nasiha-fi-sifat-rabb": true,
-  "lisan-carab": true,
-  "mukhtasar-tarikh-dimashq": true,
-  "nithar-azhar": true,
-  "tadhyil-tashil": true,
-  "tafsir-bahr-muhit": true,
-  "tuhfat-arib": true,
-  "silah-mumin": true,
-  "atwaq-hamama": true,
-  "cunwan-diraya": true,
-  "sharh-shafiya": true,
-  faiq: true,
-  "nihayat-wusul": true,
-  "dar-qawl-qabih": true,
-  intisarat: true,
-  "isharat-1": true,
-  mawaid: true,
-  "sharh-mukhtasar-rawda": true,
-  tacyin: true,
-  "ghurar-khasais": true,
-  "cawali-6": true,
-  "juz-68": true,
-  "lamha-fi-sharh-mulha": true,
-  "cunwan-dalil": true,
-  "mal-cayba": true,
-  "sunan-abyan": true,
-  ajrumiyya: true,
-  "majmac-adab": true,
-  "cudda-fi-sharh-cumda": true,
-  "ictiqad-khalis": true,
-  "tuhfat-talibin": true,
-  "dhayl-mirat-zaman": true,
-  mafatih: true,
-  "ahadith-cawali-min-ibn-carafa": true,
-  "ahadith-qussas": true,
-  "amr-bi-macruf-4": true,
-  "amrad-qulub": true,
-  "arbacuna-taymiyya": true,
-  "bayan-talbis-jahmiyya": true,
-  "bughyat-murtad": true,
-  "caqida-wasitiyya": true,
-  cubudiyya: true,
-  "daqaiq-tafsir": true,
-  "dar-tacarud": true,
-  "fadl-abi-bakr": true,
-  "fasl-fi-tazkiyyat-nafs": true,
-  "fatawa-kubra": true,
-  "fi-cabd-mahabba": true,
-  "kanz-akbar": true,
-  "sharh-tayyibat-nashr-1": true,
-  "shifa-fi-badic-iktifa": true,
-  "hudud-fi-cilm-nahw": true,
-  "barnamaj-2": true,
-  "sharh-waraqat-fi-usul-fiqh": true,
-  "majmuc-fatawa": true,
-  "majmucat-rasail-wa-masail": true,
-  "manasik-hajj": true,
-  "masala-fi-jamacat-nisa": true,
-  "masala-fi-kanais": true,
-  "masala-fi-murabata": true,
-  "masala-fi-tawhid-falasifa": true,
-  "bughyat-multamis-1": true,
-  "fusul-mufida": true,
-  "ijmal-isaba": true,
-  "itharat-fawaid": true,
-  "jamic-tahsil": true,
-  "juz-fi-tafsir-baqiyat": true,
-  "majmuc-rasail-2": true,
-  mukhtalitin: true,
-  "musalsalat-mukhtara": true,
-  "cuddat-murid": true,
-  "nasiha-kafiya": true,
-  "sharh-zarruq-cala-matn-risala": true,
-  "rawd-mictar": true,
-  "nasb-raya": true,
-  "takhrij-ahadith": true,
-  "durr-manzur": true,
-  inaba: true,
-  intikhab: true,
-  "ishara-ila-sira": true,
-  "sharh-sunan-ibn-maja": true,
-  "fawat-wafayat": true,
-  "acyan-casr": true,
-  "kashf-hal": true,
-  "lawcat-shaki": true,
-  "nakt-himyan": true,
-  "nusrat-thair": true,
-  "shucur-bi-cur": true,
-  "tashih-tashif": true,
-  "tawshic-tawshih": true,
-  "wafi-bi-wafayat": true,
-  "dhayl-tadhkira": true,
-  "ikmal-li-rijal": true,
-  "ikhtiyarat-ibn-taymiyya": true,
-  "irshad-salik-1": true,
-  "mukhtasar-kabir-1": true,
-  "hadith-77": true,
-  "diwan-134": true,
-  "marham-cilal-mucdila": true,
-  "mirat-jinan": true,
-  "akam-marjan-1": true,
-  musacid: true,
-  "sharh-cala-alfiyya": true,
-  "cudda-fi-icrab-cumda": true,
-  "misbah-munir": true,
-  "matharat-ghalat": true,
-  "miftah-wusul": true,
-  "ashbah-wa-nazair-2": true,
-  "mucid-nicam": true,
-  "mucjam-shuyukh-4": true,
-  "qacida-fi-jarh": true,
-  "qacida-fi-muarrikhin": true,
-  "rafc-hajib": true,
-  "tabaqat-shaficiyya-kubra": true,
-  "hidaya-ila-awham": true,
-  "kawkab-durri": true,
-  "matalic-daqaiq": true,
-  "muhimmat-fi-sharh-rawda": true,
-  "nihayat-sul": true,
-  "tamhid-fi-takhrij-furuc": true,
-  "sharh-cala-mukhtasar-khiraqi": true,
-  "carus-afrah": true,
-  "adab-dukhul-hammam": true,
-  "bidaya-1": true,
-  "fadail-quran-8": true,
-  "fusul-min-sira": true,
-  "jamic-masanid-2": true,
-  "mucjizat-nabi": true,
-  "musnad-cumar-ibn-khattab-2": true,
-  "nihaya-fi-fitan": true,
-  "qisas-anbiya-1": true,
-  "tabaqat-fuqaha-shaficiyyin": true,
-  takmil: true,
-  "tuhfat-talib": true,
-  "husn-suluk": true,
-  "mukhtasar-sawaciq-mursala": true,
-  "diwan-sababa": true,
-  "khasais-sayyid-calamin": true,
-  "nahj-rashad": true,
-  ihata: true,
-  "jaysh-tawshih": true,
-  "katiba-kamina": true,
-  "khatrat-tayf": true,
-  "micyar-ikhtiyar": true,
-  "nufadat-jirab": true,
-  "rayhanat-kitab": true,
-  "manhaj-qawim": true,
-  "mukhtasar-fatawa-misriyya": true,
-  "mukhtasar-sarim": true,
-  "shifa-calil-1": true,
-  "tamhid-qawacid": true,
-  "rihla-5": true,
-  kawakib: true,
-  "muqtafa-min-sirat-mustafa": true,
-  "nasim-saba": true,
-  "iqtitaf-azahir": true,
-  "tuhfat-aqran": true,
-  "manzumat-maqsur": true,
-  "tabaqat-qurra": true,
-  "misbah-mudi": true,
-  muqarrar: true,
-  "uns-mala": true,
-  "tasliyat-ahl-masaib": true,
-  "cinaya-sharh-hidaya": true,
-  "sharh-mukhtasar-ibn-hajib": true,
-  "tahqiq-fawaid": true,
-  "takhrij-dalalat": true,
-  "sifat-rabb-calamin": true,
-  ictisam: true,
-  ifadat: true,
-  muwafaqat: true,
-  "sharh-alfiyyat-ibn-malik-1": true,
-  "qawacid-wa-isharat": true,
-  "macani-badica": true,
-  "ittibac-3": true,
-  "sharh-caqida-tahawiyya-1": true,
-  "tafsir-24": true,
-  "tanbih-3": true,
-  "bahr-muhit": true,
-  "burhan-fi-culum-quran": true,
-  "iclam-sajid": true,
-  "ijaba-li-irad": true,
-  "khabaya-zawaya": true,
-  "macna-la-ilaha-illa-llah": true,
-  "manthur-fi-qawacid": true,
-  "nukat-cala-muqaddimat-ibn-salah": true,
-  "salasil-dhahab": true,
-  "tadhkira-fi-ahadith-mashhura": true,
-  "tashnif-masamic": true,
-  "zahr-carish-fi-tahrim-hashish": true,
-  "ahwal-qubur": true,
-  "akham-ikhtilaf-fi-ruya-hilal": true,
-  "asbab-maghfira": true,
-  "dhamm-khamr": true,
-  "fadl-cilm-salaf": true,
-  "farq-bayna-nasiha-wa-tacyir": true,
-  "fath-bari-fi-sharh-sahih-bukhari": true,
-  "sharh-thumam": true,
-  dalail: true,
-  "taysir-1": true,
-  "nashr-bunud": true,
-  "cajaib-athar": true,
-  "fawaid-muntakhabat": true,
-  "bulghat-salik": true,
-  "hashiya-cala-tafsir-jalalayn": true,
-  "jawab-ahl-sunna": true,
-  "kalimat-nafica": true,
-  "matalib-uli-nuha": true,
-  "ikhtiyar-ula": true,
-  "istikhraj-li-ahkam-kharaj": true,
-  "jamic-culum-wa-hikam": true,
-  "kalimat-ikhsas": true,
-  "kashf-kurba": true,
-  "lataif-macarif": true,
-  "majmuc-rasail-3": true,
-  "nur-iqtibas": true,
-  "nuzhat-asmac": true,
-  "qacida-dhahabiyya": true,
-  "qawacid-3": true,
-  "rawaic-tafsir-fatiha": true,
-  "sharh-cilal-tirmidhi": true,
-  "sharh-hadith-labbayka": true,
-  "sharh-hadith-yatbac-mayyit-thalath": true,
-  "takhwif-min-nar": true,
-  "tasliyat-nufus-nisa": true,
-  "farq-bayn-dad-wa-zad": true,
-  rasf: true,
-  "mahasin-istilah": true,
-  "dibaj-mudhahhab": true,
-  "irshad-salik-2": true,
-  "tabsirat-hukkam": true,
-  "ahadith-17": true,
-  "shicar-abrar": true,
-  "ishara-ila-cilm-cibara": true,
-  "jawhara-nayyira": true,
-  "siraj-qari": true,
-  "shadha-fayyah": true,
-  "mukhtasar-fi-mantiq": true,
-  "mukhtasar-fiqhi": true,
-  "tafsir-25": true,
-  "mukhtasar-fi-usul-fiqh": true,
-  "qawacid-4": true,
-  "man-takallama-fihi-daraqutni": true,
-  "kashf-manahij": true,
-  ashbah: true,
-  "badr-munir": true,
-  "ciqd-mudhhab": true,
-  "cujalat-muhtaj": true,
-  "ghayat-sul": true,
-  "iclam-bi-fawaid": true,
-  "khulasat-badr-munir": true,
-  mucin: true,
-  "muqnic-fi-culum-hadith": true,
-  "tabaqat-awliya": true,
-  "tadhkira-fi-culum-hadith": true,
-  "tadhkira-fi-fiqh-shafici": true,
-  "tadhkirat-muhtaj": true,
-  "tawdih-2": true,
-  "tuhfat-muhtaj": true,
-  "shamil-fi-fiqh-malik": true,
-  "tahbir-mukhtasar": true,
-  "fawaid-73": true,
-  "tadrib-fi-fiqh-shafici": true,
-  "alfiyyat-sira": true,
-  "alfiyyat-tabsira": true,
-  "arbacuna-cushariyya": true,
-  "dhayl-mizan": true,
-  "khamsat-ahadith": true,
-  "mustakhraj-cala-mustadrak": true,
-  "qurrat-cain": true,
-  taqyid: true,
-  "tarh-tarthib": true,
-  "tawsica-cala-ciyal": true,
-  "aclam-maghrib": true,
-  "nafha-nisriniyya": true,
-  "sharh-cala-alfiyya-1": true,
-  "kashf-astar": true,
-  "majmac-zawaid": true,
-  "maqsad-cali": true,
-  "mawarid-zaman": true,
-  "adab-akl": true,
-  "rihla-6": true,
-  "tarikh-10": true,
-  "hayat-hayawan-kubra": true,
-  "najm-wahhaj": true,
-  "sharh-lamiyyat-cajam": true,
-  "nuzhat-anam": true,
-  "cuqud-luluiyya": true,
-  "tanbih-ghafilin-2": true,
-  "matalic-budur": true,
-  "tibyan-fi-tafsir-gharib": true,
-  "arbacuna-min-macali": true,
-  "mashyakha-28": true,
-  "tarikh-11": true,
-  "dibaj-mudhahhab-1": true,
-  nahw: true,
-  tacrifat: true,
-  "basair-dhawi-tamyiz": true,
-  "bulgha-fi-tarajim": true,
-  maghanim: true,
-  "qamus-muhit": true,
-  "radd-cala-rafida-1": true,
-  "risala-fi-bayan": true,
-  "safar-sacada-1": true,
-  "tanwir-miqbas": true,
-  "ghurar-wa-durar": true,
-  maathir: true,
-  "nihayat-arab-1": true,
-  "qalaid-juman": true,
-  "subh-acsha": true,
-  "taysir-bayan": true,
-  "ghayth-hamic": true,
-  mudallisin: true,
-  "tahrir-fatawa": true,
-  "tuhfat-tahsil": true,
-  "cuyun-ghamiza": true,
-  "masabih-jamic": true,
-  "tuhfat-hukkam": true,
-  "kifayat-akhyar": true,
-  "fawaid-saniyya": true,
-  "tuhfat-rakic": true,
-  "kunuz-dhahab": true,
-  "maqsid-arshad": true,
-  "mubdic-fi-sharh-muqnic": true,
-  "masacid-nazar": true,
-  "masrac-tasawwuf": true,
-  "nazm-durar-1": true,
-  "nukat-wafiyya": true,
-  "insaf-4": true,
-  "tahbir-sharh-tahrir": true,
-  "tahrir-manqul": true,
-  "durar-hukkam": true,
-  "sharh-shudhur-dhahab-1": true,
-  "bahjat-mahafil": true,
-  "sharh-alfiyyat-ciraqi": true,
-  "durar-lawamic": true,
-  "ghayat-amani": true,
-  "kawthar-jari": true,
-  "ahadith-qudsiyya": true,
-  "asrar-marfuca": true,
-  "fath-bab-cinaya": true,
-  "fayd-mucin": true,
-  "jamc-wasail": true,
-  "mawcizat-habib": true,
-  "mawducat-sughra": true,
-  "mirqat-mafatih": true,
-  "radd-cala-qailin-bi-wahdat-wujud": true,
-  "shamm-cawarid": true,
-  "sharh-musnad-abi-hanifa": true,
-  "sharh-nukhbat-fikr": true,
-  "ifadat-mubtadi": true,
-  "risala-4": true,
-  "tacliq-rashiq": true,
-  "sharh-cala-alfiyyat-malik": true,
-  "ajwiba-murdiyya": true,
-  buldaniyyat: true,
-  "daw-lamic": true,
-  "fakhr-mutawali": true,
-  "fath-mughith": true,
-  "ghaya-fi-sharh": true,
-  "iclan-bi-tawbikh": true,
-  "iltimas-sacd": true,
-  intihad: true,
-  "jawahir-wa-durar": true,
-  "manhal-cadhb": true,
-  "maqasid-hasana": true,
-  "mutakallimun-fi-rijal": true,
-  "qanaca-1": true,
-  "qawl-badic": true,
-  "sirr-maktum": true,
-  "tajrid-cala-tanqih": true,
-  "tawdih-abhar": true,
-  "tuhfa-latifa": true,
-  "tabaqat-sulaha-yaman": true,
-  "tarikh-12": true,
-  "muwassil-tullab": true,
-  "sharh-azhariyya": true,
-  "sharh-tasrih": true,
-  "tafsir-27": true,
-  "arbacuna-min-hadith-abi-hanifa": true,
-  "arbacuna-musalsala": true,
-  "bahr-damm": true,
-  "cashara-min-marwiyyat-salih": true,
-  "ciqd-tamam": true,
-  dabt: true,
-  "durr-naqi": true,
-  "ghayat-sul-1": true,
-  "hisba-1": true,
-  "idah-maqala": true,
-  "idah-turuq-istiqama": true,
-  "ikhtilaf-bayna-ruwat-bukhari": true,
-  ikhtiyar: true,
-  "isticana-cala-fatiha": true,
-  "jamc-juyush-wa-dasakir": true,
-  "jawab-bacd-khadam": true,
-  jawhar: true,
-  "macarif-incam": true,
-  "mahd-sawab": true,
-  "mucjam-kutub": true,
-  "nihaya-1": true,
-  nujat: true,
-  "sabb-khumul": true,
-  "siyar-hathth": true,
-  "tadhkirat-huffaz-2": true,
-  "takhrij-saghir-wa-tahbir-kabir": true,
-  "zinat-carais": true,
-  "zubad-culum": true,
-  "asrar-kawn": true,
-  "bahr-alladhi-zakhar": true,
-  "bast-kaff": true,
-  "bughyat-wucat": true,
-  "bushra-kaib": true,
-  "buzugh-hilal": true,
-  "jamic-kabir-1": true,
-  "jamic-saghir-1": true,
-  "jiyad-musalsalat": true,
-  "juhd-qariha": true,
-  "khasais-kubra": true,
-  "laali-masnuca": true,
-  "lubab-hadith": true,
-  "lubab-nuqul": true,
-  "lubb-lubab": true,
-  "lumac-fi-asbab-hadith": true,
-  "lumca-fi-tahqiq": true,
-  "madraj-ila-mudraj": true,
-  "manahil-safa": true,
-  "maqamat-4": true,
-  "marasid-matalic": true,
-  "matlac-badrayn": true,
-  "miftah-janna": true,
-  "mimma-rawahu-asatin": true,
-  "mirqat-sucud": true,
-  "misbah-zujaja-1": true,
-  "mucjam-maqalid": true,
-  "muctarak-aqran": true,
-  "mufhimat-aqran": true,
-  muhadarat: true,
-  "muhadhdhab-fima-waqaca": true,
-  "muzhir-fi-culum-lugha": true,
-  "nawahid-abkar": true,
-  "nazm-ciqyan": true,
-  "nuzhat-julasa": true,
-  "fasl-khitab-1": true,
-  "ghayat-amani-1": true,
-  "ma-dalla-calayhi-quran": true,
-  "mukhtasar-tuhfa-ithna-cashariyya": true,
-  "sabb-cadhab": true,
-  suyuf: true,
-  "tahdhir-khawass": true,
-  "takhir-zalama": true,
-  "takrir-istinad": true,
-  "tamhid-farsh": true,
-  "tanwir-hawalik": true,
-  "tarikh-khulafa": true,
-  tatrif: true,
-  "tawq-hamama-1": true,
-  tawshih: true,
-  "tiraz-fi-alghaz": true,
-  "tuhfat-abrar-1": true,
-  "unmudhaj-labib": true,
-  "ziyadat-cala-mawducat": true,
-  "asna-matajir": true,
-  "cuddat-buruq": true,
-  "idah-masalik": true,
-  "husam-maslul": true,
-  "muqaddima-hadramiyya": true,
-  "hujaj-bahira": true,
-  "fath-qarib": true,
-  "shifa-ghalil-1": true,
-  "ghayat-sul-2": true,
-  "nayl-amal": true,
-  "fawatih-ilahiyya": true,
-  "iscaf-awqaf": true,
-  "mawahib-laduniyya": true,
-  "tahrir-suluk": true,
-  "asna-matalib-1": true,
-  "fath-baqi": true,
-  "fath-callam": true,
-  "fath-rahman": true,
-  "mukaffirat-dhunub": true,
-  "fath-wahhab": true,
-  "ghayat-wusul": true,
-  "ghurar-bahiyya": true,
-  "hudud-aniqa": true,
-  "icrab-quran-4": true,
-  "manhaj-tullab": true,
-  "maqsid-li-talkhis": true,
-  "minhat-bari": true,
-  munfarijatan: true,
-  "daris-fi-madaris": true,
-  "kanz-akbar-1": true,
-  "tarikh-muctabar": true,
-  "kawakib-nayyirat": true,
-  "fath-aqfal": true,
-  "hadaiq-anwar": true,
-  "tabaqat-mufassirin-1": true,
-  "qiladat-nahr": true,
-  "sharh-qawacid-icrab": true,
-  "ahadith-mia-2": true,
-  "fass-khawatim": true,
-  "iclam-sailin": true,
-  "inba-umara": true,
-  "mufakaha-khillan": true,
-  "tashyid-ikhtiyar": true,
-  "mawahib-jalil": true,
-  "mutammimat-ajrumiyya": true,
-  "tahrir-kalam": true,
-  "nicmat-dharica": true,
-  "kafil-bi-nayl-sul": true,
-  "fatawa-ramli": true,
-  "fath-rahman-2": true,
-  "macahid-tansis": true,
-  "tanzih-sharica": true,
-  "iqnac-2": true,
-  "zad-mustaqnic": true,
-  "shaqaiq-nucmaniyya": true,
-  "ashbah-wa-nazair-3": true,
-  "bahr-raiq": true,
-  "qafw-athar": true,
-  "sahm-alhaz": true,
-  "muntaha-iradat": true,
-  "sharh-kawkab-munir": true,
-  "taysir-tahrir": true,
-  "ashraf-wasail": true,
-  "durr-mandud-1": true,
-  "fatawa-hadithiyya": true,
-  "fatawa-kubra-1": true,
-  "fath-mubin": true,
-  "iclam-3": true,
-  "ifsah-can-hadith-nikah": true,
-  inafa: true,
-  "kaff-ricac": true,
-  "mablagh-arab": true,
-  "manhaj-qawim-1": true,
-  "sawaciq-muhriqa": true,
-  "tuhfat-muhtaj-1": true,
-  zawajir: true,
-  "kanz-cummal": true,
-  "iqnac-3": true,
-  "khisal-mukaffira-1": true,
-  "mughni-muhtaj": true,
-  "siraj-munir": true,
-  "anis-fuqaha": true,
-  "khulasat-jawahir-zakiyya": true,
-  "ciqd-talid": true,
-  "risalat-inqadh-halikin": true,
-  "irshad-caql-salim": true,
-  "mukhtasar-fi-cibadat": true,
-  "adab-cishra": true,
-  "adab-muwakala": true,
-  "matalic-badriya": true,
-  "murah-fi-muzah": true,
-  "majmac-bihar": true,
-  "tadhkirat-mawducat-1": true,
-  "fath-mucin": true,
-  "isticdad-li-mawt": true,
-  "durra-bahiyya": true,
-  "nihayat-tadrib": true,
-  "sharh-manhaj-muntakhab": true,
-  "thalathat-usul": true,
-  "usul-din-3": true,
-  "usul-iman": true,
-  "hashiya-cala-sharh-ashmuni": true,
-  "silk-durar": true,
-  "minhat-qarib-mujib": true,
-  "fulk-qamus": true,
-  "risala-fi-hisba": true,
-  "risala-muhimma": true,
-  "iqaz-himam": true,
-  "qatf-thamar": true,
-  "fath-shakur": true,
-  "tuhfat-mushtaq": true,
-  "tajrid-li-nafc-cabid": true,
-  "tuhfat-habib": true,
-  "bahr-madid": true,
-  "ciqd-thamin-3": true,
-  "fawakih-cidhab": true,
-  "majmucat-rasail-2": true,
-  "rasail-fiqhiyya-1": true,
-  "rasail-wa-fatawa-1": true,
-  "tuhfa-madaniyya": true,
-  "tafsir-mazhari": true,
-  fasiha: true,
-  mufakhara: true,
-  "nuzhat-anzar": true,
-  "nukhba-bahiyya": true,
-  "hashiya-cala-mukhtasar-macani": true,
-  "hashiyat-cala-sharh-kabir": true,
-  "hashiya-cala-maraqi": true,
-  "risalat-tawhid-1": true,
-  "durar-mafakhir": true,
-  "hashiya-cala-jamc-jawamic": true,
-  "tanqih-fatawa": true,
-  "nafhat-yaman": true,
-  mirat: true,
-  "intikhab-cawali": true,
-  "suyuf-bawatir": true,
-  mudhakkirat: true,
-  "sharh-muqaddima-hadramiyya": true,
-  "tanbih-afham": true,
-  "ajwiba-ciraqiyya": true,
-  "gharaib-ightirab": true,
-  "ruh-macani": true,
-  "irshad-cibad": true,
-  "fath-ghaffar": true,
-  "asna-matalib-2": true,
-  "dhayl-qawl-musaddad": true,
-  "fi-suluk-islami": true,
-  "mawrid-cadhb-zulal": true,
-  "rasail-wa-fatawa-3": true,
-  "tawhid-wa-qurrat-cuyun-muwahhidin": true,
-  "majmac-bahrayn-1": true,
-  "nihayat-ijaz": true,
-  "matalic-nasriyya": true,
-  "qurrat-cayn-bi-fatawa": true,
-  "barahin-islamiyya": true,
-  "cuyun-rasailwa-ajwiba": true,
-  "ijaza-2": true,
-  "ithaf-1": true,
-  "minhaj-tasis-wa-taqdis": true,
-  "misbah-zalam": true,
-  "tuhfat-talib-1": true,
-  "lubab-fi-sharh-kitab": true,
-  "hamasa-2": true,
-  "nukat-shanica": true,
-  "fath-cali": true,
-  "minah-jalil": true,
-  "takmilat-macajim-carabiyya": true,
-  "lawamic-durar": true,
-  "jasus-cala-qamus": true,
-  "saq-cala-saq": true,
-  "athar-marfuca": true,
-  "fawaid-bahiyya": true,
-  "insaf-fi-hukm-ictikaf": true,
-  "radc-ikhwan": true,
-  "rafc-wa-takmil": true,
-  "tacliq-mumajjad": true,
-  "ciqd-mufassal": true,
-  "lulu-marsuc": true,
-  "nazm-laal": true,
-  "nubdha-latifa": true,
-  "murshid-hayran": true,
-  "abjad-culum": true,
-  "bulgha-ila-usul-lugha": true,
-  "fath-bayan": true,
-  "hitta-fi-dhikr": true,
-  "husn-uswa": true,
-  "luqtat-cajlan": true,
-  "nashwat-sakran": true,
-  "nayl-maram": true,
-  "qatf-thamar-1": true,
-  "rawda-nadiyya": true,
-  "rihla-9": true,
-  "taj-mukallal": true,
-  yaqza: true,
-  "izhar-haqq": true,
-  "mukhtasar-izhar-haqq": true,
-  "munazara-taqririyya": true,
-  "hashiya-icana-talibin": true,
-  "iktifa-qunuc": true,
-  "majallat-tankit": true,
-  "majallat-ustadh": true,
-  "nafh-azhar": true,
-  "istiqsa-li-ahkbar": true,
-  "ayat-bayyinat-1": true,
-  "jala-caynayn": true,
-  musamarat: true,
-  "hashiyat-cala-nayl-maarib": true,
-  "rasail-bushra": true,
-  "kawkab-durri-2": true,
-  "nujcat-raid": true,
-  "nujcat-raid-1": true,
-  "siyanat-insan": true,
-  "radd-cala-shubuhat": true,
-  "tawdih-maqasid-2": true,
-  "cawn-macbud": true,
-  "nadrat-bahar": true,
-  "wasit-fi-tarajim": true,
-  "nukhbat-ciqd": true,
-  "islah-masajid": true,
-  "jarh-wa-tacdil-1": true,
-  "mawcizat-muminin": true,
-  "qawacid-tahdith": true,
-  "risala-fi-shay": true,
-  "sirr-istighfar": true,
-  "thamarat-tasaruc": true,
-  "durr-manthur-1": true,
-  "hilyat-bashar": true,
-  "thamr-dani": true,
-  "mudhakkarati-siyasiya": true,
-  "nuzhat-afkar": true,
-  "tarikh-13": true,
-  "tawjih-nazar": true,
-  "hadiyyat-carifin": true,
-  "majani-adab": true,
-  nasraniyya: true,
-  "tarikh-adab-carabiyya": true,
-  "awham-shucara-carab": true,
-  "mucjam-taymur-kabir": true,
-  "nazra-tarikhiyya": true,
-  "samac-wa-qiyas": true,
-  "tashih-lisan-carab": true,
-  "minhaj-ahl-haqq": true,
-  "nazam-ikhtiyarat": true,
-  "nazam-ma-infarada-bihi": true,
-  "sawaciq-mursala-1": true,
-  "tamayyuz-sidq": true,
-  "tanbih-dhawi-albab": true,
-  "dalil-hayran": true,
-  "adab-nabawi": true,
-  "nahr-dhahab": true,
-  "mucjam-matbucat": true,
-  "tuhfat-ahwadhi": true,
-  khilafa: true,
-  manar: true,
-  "wahy-muhammadi": true,
-  "tanbih-wa-iqaz": true,
-  "icjaz-quran-1": true,
-  masakin: true,
-  "sumuww-ruhi": true,
-  "tahta-rayat-quran": true,
-  "tarikh-adab-carab": true,
-  "wahy-qalam": true,
-  "tarikh-bimaristanat-fi-islam": true,
-  "jawahir-adab": true,
-  "jawahir-balagha": true,
-  "majallat-haqaiq": true,
-  "nuzhat-absar-1": true,
-  "dhakhair-1": true,
-  "majallat-bayan": true,
-  "muntakhab-qabail": true,
-  "majallat-lughat-carab": true,
-  "nahw-wadih": true,
-  "abu-bakr": true,
-  cuthman: true,
-  muhammad: true,
-  "tafsir-31": true,
-  "khitat-sham": true,
-  "majallat-muqtabas": true,
-  "majallat-thaqafa": true,
-  "rihla-shamiyya": true,
-  "taysir-karim-rahman": true,
-  "fikr-sami": true,
-  "bacith-hathith": true,
-  "jamharat-maqalat": true,
-  "majmuc-fatawa-2": true,
-  "majmucat-qasaid": true,
-  "rijal-hakim": true,
-  "mucjam-huffaz-quran": true,
-  "tarajim-rijal": true,
-  "tarikh-falsafa-haditha": true,
-  "tarikh-falsafa-urubbiyya-fi-casr-wasit": true,
-  "tarikh-falsafa-yunaniyya": true,
-  "isharat-icjaz": true,
-  "ajwiba-kafiya": true,
-  "fatawa-wa-rasail": true,
-  "iscaf-acyan": true,
-  "adwa-bayan": true,
-  "jamharat-maqalat-1": true,
-  tahqiqat: true,
-  "tahrir-wa-tanwir": true,
-  "khatim-nabiyyin": true,
-  aclam: true,
-  "ithaf-matalic": true,
-  "taysir-tafsir-1": true,
-  mashahir: true,
-  "mucjam-muallifin": true,
-  "mucjam-qabail": true,
-  "mufassal-fi-tarikh-carab": true,
-  "nubugh-maghribi": true,
-  "hidayat-qari": true,
-  "tacliqa-cala-fawaid-radawiyya": true,
-  "mudhakkirat-tawhid": true,
-  "min-tarikh-nahw-carabi": true,
-  "tuhfat-majalis": true,
-  "maqalat-tanahi": true,
-  "mujiz-fi-marajic": true,
-  "madha-khasara-calam": true,
-  "fi-sabil-islah": true,
-  "fikar-wa-mabahith": true,
-  "fusul-fi-dacwa": true,
-  "fusul-fi-thaqafa": true,
-  "min-hadith-nafs": true,
-  "nur-wa-hidaya": true,
-  "suwar-wa-khawatir": true,
-  "tacrif-camm": true,
-  "fatawa-nur-cala-darb": true,
-  "mucjam-qabail-mcs": true,
-  "fatawa-arkan-islam": true,
-  "fatawa-nur-cala-darb-1": true,
-  "tadhkira-uli-nahy": true,
-  "tabaqat-nassabin": true,
-  "fatawa-6": true,
-  "manahij-mufassirin": true,
-  "tafsir-wasit": true,
-  "aysar-tafasir": true,
-  "muntakhab-min-cilal": true,
-  "muqnic-fi-fiqh": true,
-  "mutahabbin-fi-llah": true,
-  "rawdat-nazir": true,
-  "riqqa-wa-buka-1": true,
-  "risala-fi-quran": true,
-  "tahrim-nazar": true,
-  tawwabin: true,
-  "mughni-can-hifz-wa-kitab": true,
-  "majmuc-takhrij": true,
-  "cudda-sharh-cumda": true,
-  "hadith-cisa-ibn-maryam": true,
-  "iqtidab-1": true,
-  "nihayat-rutba-1": true,
-  "miftah-culum": true,
-  khazal: true,
-  "mucjam-buldan": true,
-  "isaba-fi-tamyiz": true,
-  "mucjam-udaba": true,
-  "bayan-wahm": true,
-  "ihkam-naza": true,
-  "iqnac-1": true,
-  "ifada-wa-ictibar": true,
-  "takmilat-ikmal": true,
-  "taqyid-li-macrifa": true,
-  kamil: true,
-  "lubab-fi-tahdhib-ansab": true,
-  "usd-ghaba": true,
-  "tahdhib-riyasa": true,
-  "cawali-malik-6": true,
-  "ghayat-maram": true,
-  "ihkam-fi-usul-ahkam-1": true,
-  "nawadir-sultaniyya": true,
-  "nazm-mustaczab": true,
-  "ada-ma-wajaba": true,
-  "ma-wadaha-wa-ma-istabana": true,
-  "mutrib-min-ashcar": true,
-  "macrifat-farq": true,
-  iktifa: true,
-  "musalsalat-2": true,
-  "istikhraj-jidal": true,
-  "mashyakha-15": true,
-  "iclam-2": true,
-  "nasiha-li-raci": true,
-  muclim: true,
-  "cubab-zakhir": true,
-  "mawducat-1": true,
-  shawarid: true,
-  "takmila-2": true,
-  "mashyakha-baghdadiyya-1": true,
-  "kanz-kuttab": true,
-  muharrar: true,
-  "tahrir-tahbir": true,
-  qalaid: true,
-  "mukhtasar-sana-barq-shami": true,
-  "juz-63": true,
-  "durra-thamina": true,
-  "adab-mufti-wa-mustafti": true,
-  "fatawa-1": true,
-  "muqaddimat-culum-hadith": true,
-  "sharh-mushkil-wajiz": true,
-  "siyanat-sahih-muslim": true,
-  "tabaqat-fuqaha-1": true,
-  "wasl-balaghat-muwatta": true,
-  "sharh-mufassal": true,
-  "farid-fi-icrab": true,
-  "masala-mardiniyya": true,
-  "minhaj-sunna": true,
-  "muqaddima-fi-usul-tafsir": true,
-  "mustadrak-cala-majmuc-fatawa": true,
-  "naqd-maratib-ijmac": true,
-  nubuwwat: true,
-  nusayriyya: true,
-  "qacida-cazima-fi-farq": true,
-  "qacida-fi-amwal-sultaniyya": true,
-  "qacida-fi-inghimas": true,
-  "qacida-fi-mahabba": true,
-  "qacida-fi-qatl-kuffar": true,
-  "qacida-fi-sabr": true,
-  "qacida-hasana-fi-baqiyat": true,
-  "qacida-jalila": true,
-  "qacida-jamica-fi-tawhid": true,
-  "qacida-mukhtacira-fi-wujub-taca": true,
-  "qacida-tatadammanu-dhikr-malabis-nabi": true,
-  "qasida-taiyya-fi-qadar": true,
-  "qawacid-nuraniyya": true,
-  "radd-cala-akhnai": true,
-  "radd-cala-man-qala-bi-fana-janna": true,
-  "radd-cala-mantiqiyyin": true,
-  "radd-cala-shadhili": true,
-  "radd-cala-subki": true,
-  "rafc-malam": true,
-  "ras-husayn": true,
-  "risala-akmaliyya": true,
-  "risala-carshiyya": true,
-  "risala-fi-cilm-qiraat": true,
-  "risala-fi-dukhul-janna": true,
-  "risala-fi-fadl-khulafa-rashidin": true,
-  "risala-fi-lafz-sunna": true,
-  "risala-fi-macna-kawn-rabb": true,
-  "risala-fi-qunut-ashya": true,
-  "risala-fi-radd-cala-ibn-carabi": true,
-  "risala-fi-tahqiq-shukr": true,
-  "risala-fi-tahqiq-tawakkul": true,
-  "risala-fi-usul-din": true,
-  "risala-fi-wa-istacinu-bi-sabr": true,
-  "risala-madaniyya": true,
-  "risala-qubrusiyya": true,
-  "risala-tadmuriyya": true,
-  safadiyya: true,
-  "sarim-maslul": true,
-  "sharh-caqida-isfahaniyya": true,
-  "sharh-cumdat-fiqh": true,
-  "sharh-hadith-nuzul": true,
-  "siyasa-sharciyya": true,
-  "sujud-tilawa": true,
-  "sunnat-jumca": true,
-  "tahqiq-qawl-fi-masalat-cisa": true,
-  "tanbih-rajul-caqil": true,
-  "tawba-2": true,
-  tisciniyya: true,
-  "tuhfa-ciraqiyya": true,
-  "wasita-bayna-haqq-wa-khalq": true,
-  "kashshaf-qinac": true,
-  "minah-shafiyat": true,
-  "rawd-murbic": true,
-  "sharh-muntaha-iradat": true,
-  "lumcat-tanqih": true,
-  "muqaddima-fi-usul-hadith": true,
-  "dalil-falihin": true,
-  "futuhat-rabbaniyya": true,
-  "ithaf-fadil": true,
-  "nashr-alwiyat-tashrif": true,
-  "qawl-sadid": true,
-  "husn-tanabbuh": true,
-  "kawakib-saira": true,
-  "ziyarat-qubur": true,
-  "zuhd-wa-warac": true,
-  "macalim-qurba": true,
-  "kashf-asrar-1": true,
-  barnamaj: true,
-  "fadl-mathur": true,
-  "kunnash-fi-nahw-wa-sarf": true,
-  "mukhtasar-fi-akhbar": true,
-  yawaqit: true,
-  "rusukh-ahbar": true,
-  "rusum-tahdith": true,
-  "irshad-salik": true,
-  "suluk-fi-tabaqat": true,
-  "wajiz-fi-fiqh": true,
-  "ahadith-tusaciyya": true,
-  "idah-dalil": true,
-  "kashf-macani": true,
-  "sharh-mishkat": true,
-  ikhtiyarat: true,
-  "majmuc-rasail-1": true,
-  "muharrir-fi-hadith": true,
-  "sarim-munki": true,
-  "tabaqat-culama-1": true,
-  "tacliqa-cala-cilal": true,
-  "tanqih-tahqiq": true,
-  irtishaf: true,
-  "ghayat-maqsad": true,
-  "tiraz-li-asrar": true,
-  "talic-sacid": true,
-  "tahdhib-sunan-abi-dawud": true,
-  "bayan-mukhtasar": true,
-  "jana-dani": true,
-  "tawdih-maqasid": true,
-  "nukhab-dhakhair": true,
-  "aclam-caliyya": true,
-  mustafad: true,
-  "masalik-absar": true,
-  "tuhfat-masul": true,
-  "tacrif-bi-mustalah-sharif": true,
-  "ahadith-sihah": true,
-  munawwar: true,
-  "sharh-alfiyyat-ibn-malik": true,
-  "tarikh-9": true,
-  "mashyakha-26": true,
-  "rahmat-calamin": true,
-  "mufradat-quran": true,
-  "bayan-mubdi": true,
-  "diya-shariq": true,
-  futyayan: true,
-  "iqamat-hujja": true,
-  "kashf-awham": true,
-  "kashf-ghayahib": true,
-  "kashf-shubhatayn": true,
-  "aclam-muwaqqicin": true,
-  "ahkam-ahl-dhimma": true,
-  "amthal-fi-quran": true,
-  "asma-muallafat-ibn-taymiya": true,
-  "badaic-fawaid": true,
-  "cuddat-sabirin": true,
-  "da-wa-dawa": true,
-  "faida-jalila": true,
-  furusiyya: true,
-  "futya-fi-sighat-hamd": true,
-  "hadi-arwah": true,
-  "hidayat-hayara": true,
-  "ighathat-lahfan": true,
-  "ijtimac-juyush": true,
-  "jala-afham": true,
-  "jawab-fi-siyagh-hamd": true,
-  "kalam-cala-masalat-samac": true,
-  "manar-munif": true,
-  "miftah-dar-sacada": true,
-  "rafc-yadayn-fi-salat": true,
-  "rawdat-muhibbin": true,
-  "risala-ila-ahad-ikhwan": true,
-  "risala-tabukiyya": true,
-  "sawaciq-mursala": true,
-  "shifa-calil": true,
-  "sifat-munafiqin": true,
-  "tibb-nabawi-2": true,
-  "tibyan-fi-aqsam-quran": true,
-  "tuhfat-mawdud": true,
-  "turuq-hukmiyya": true,
-  "wabil-sayyib": true,
-  "zad-macad": true,
-  "sharh-cadud": true,
-  "cumdat-huffaz": true,
-  "durr-masun": true,
-  "fatawa-2": true,
-  "ibraz-hikam": true,
-  "qada-arab": true,
-  "sayf-maslul": true,
-  "wasiyya-3": true,
-  "musnad-17": true,
-  "tuhfat-turk": true,
-  lamic: true,
-  "ciqd-thamin-2": true,
-  "dhayl-taqyid": true,
-  "shifa-gharam": true,
-  "durra-mudiyya-1": true,
-  "ghayat-nihaya": true,
-  "manaqib-asad-ghalib": true,
-  "matn-tayyibat-nashr": true,
-  "munjid-muqriyyin": true,
-  nashr: true,
-  "tahbir-taysir": true,
-  "tamhid-fi-culum-tajwid": true,
-  "zahr-faih": true,
-  "fusul-badaic": true,
-  "bulugh-amal": true,
-  falaka: true,
-  "cawasim-wa-qawasim": true,
-  "ithar-haqq": true,
-  "rawd-basim": true,
-  "ithaf-khayra": true,
-  "misbah-zujaja": true,
-  ightibat: true,
-  "kashf-hathith": true,
-  "tabyin-li-asma": true,
-  "ahadith-arbacuna": true,
-  "ahadith-sitta": true,
-  "carf-canbar": true,
-  "imla-anfus": true,
-  "majlis-fi-hadith-jabir": true,
-  "salwat-kaib": true,
-  "tanwir-fikra": true,
-  "tawdih-mushtabih": true,
-  "durra-gharra": true,
-  "sharh-sunan-abi-dawud": true,
-  "zubad-fi-fiqh": true,
-  "mucin-hukkam": true,
-  "miftah-sacidiyya": true,
-  "bayan-2": true,
-  "daw-sari": true,
-  "imtac-asmac": true,
-  "itticaz-hunafa": true,
-  "mawaciz-1": true,
-  "mukhtasar-kamil": true,
-  "mukhtasar-qiyam-layl": true,
-  "mukhtasar-qiyam-ramadan": true,
-  muqaffa: true,
-  "rasail-6": true,
-  suluk: true,
-  "tajrid-tawhid": true,
-  "fatawa-4": true,
-  "tabaqat-shaficiyya-1": true,
-  mukarrar: true,
-  "rawd-akhyar": true,
-  "risala-fi-tahqiq-macna": true,
-  "talwin-khitab": true,
-  "subul-huda": true,
-  "jawahir-durar": true,
-  "bughrat-irba": true,
-  "ahadith-cashara": true,
-  "amali-adhkar-fi-fadl-salat-tasbih": true,
-  "amali-halabiyya": true,
-  "amali-mutlaqa": true,
-  "bulugh-maram": true,
-  "cujab-fi-bayan-asbab": true,
-  "diraya-fi-takhrij-ahadith-hidaya": true,
-  "durar-kamina": true,
-  "fath-bari": true,
-  "imtac-bi-arbacin": true,
-  "inba-ghumr": true,
-  "intiqad-ictirad": true,
-  "ithar-bi-macrifa": true,
-  "itraf-musnad-muctali": true,
-  "kalam-cala-hadith-imraati": true,
-  "lisan-mizan": true,
-  "majmac-muassis": true,
-  "marhama-ghaythiyya": true,
-  "masail-ajaba-canha": true,
-  "matalib-caliya-1": true,
-  "mucjam-shaykha-maryam": true,
-  "mukhtasar-zawaid-musnad-bazzar": true,
-  "muntakhab-min-amali": true,
-  "muwafaqat-khubr-khabar": true,
-  "nazm-laali": true,
-  "nukat-cala-ibn-salah": true,
-  "nukat-cala-sahih-bukhari": true,
-  "nukhbat-fikar": true,
-  "nuzhat-albab-1": true,
-  "nuzhat-nazar": true,
-  "nuzhat-samicin": true,
-  "qawl-musaddad": true,
-  "rafc-isr": true,
-  "silsilat-dhahab": true,
-  "tabsir-muntabih": true,
-  "tacjil-manfaca": true,
-  "taghliq-tacliq": true,
-  "tahdhib-tahdhib": true,
-  "takhrij-ahadith-asma-husna": true,
-  "talkhis-habir": true,
-  "taqrib-tahdhib": true,
-  "turuq-hadith-la-tasubbu-ashabi": true,
-  "wuquf-cala-mawquf": true,
-  "zahr-nadir-fi-hal-khadir": true,
-  "kharidat-cajaib": true,
-  "cajaib-maqdur": true,
-  "fakihat-khulafa": true,
-  "juz-74": true,
-  "mukhtasar-tanzih-masjid": true,
-  "tarikh-makka": true,
-  "sharh-masabih": true,
-  "bahjat-nazirin": true,
-  "ramuz-cala-sihah": true,
-  "tuhfat-amin": true,
-  "bisharat-mahbub": true,
-  "anjum-zahirat": true,
-  "tuhfat-nazir": true,
-  "lahz-alhaz": true,
-  "isharat-2": true,
-  "bidayat-muhtaj": true,
-  "taysir-wusul-1": true,
-  "hawadith-duhur": true,
-  "manhal-safi": true,
-  "mawrid-latafa": true,
-  "nujum-zahira": true,
-  "jannat-wildan": true,
-  "kans-jawari": true,
-  "taqrir-wa-tahbir": true,
-  "ahkam-salat-cala-janaza": true,
-  "khulasat-afkar": true,
-  "majmuc-rasail-4": true,
-  "musnad-cuqba": true,
-  "rafc-ishtibah": true,
-  "tahrir-aqwal": true,
-  "taj-tarajim-1": true,
-  "thiqat-1": true,
-  "mukhtasar-fi-cilm-athar": true,
-  "ithaf-akhissa": true,
-  "jawahir-cuqud": true,
-  "farah-asma": true,
-  "lisan-hukkam": true,
-  "sharh-mukhtasar-usul-fiqh": true,
-  "ghayat-bayan": true,
-  "nihayat-muhtaj": true,
-  "badhl-majhud": true,
-  "nahr-faiq": true,
-  "tazyin-aswaq": true,
-  "tabaqat-saniyya": true,
-  adilla: true,
-  "taj-carus": true,
-  "takhrij-ahadith-ihya-culum-din": true,
-  "sharh-shifa": true,
-  "tasliyyat-acma": true,
-  "shamca-mudiyya": true,
-  "hadi-azcan": true,
-  "durar-faraid": true,
-  "durrat-hijal": true,
-  "majmac-damanat": true,
-  kashkul: true,
-  "fath-samawi": true,
-  "fayd-qadir": true,
-  ittihafat: true,
-  "sayyidat-nisa": true,
-  "sharh-arbacin-nawawiyya": true,
-  tawqif: true,
-  "taysir-bi-sharh": true,
-  "yawaqit-wa-durar": true,
-  "aqawil-thiqat": true,
-  "dalil-talib-cala-madhhab-ibn-hanbal": true,
-  "dalil-talib-li-nayl-matalib": true,
-  "dalil-talibin-li-kalam-nahwiyyin": true,
-  "fawaid-mawduca": true,
-  "ghayat-muntaha": true,
-  "ghidha-arwah": true,
-  "ikhlas-widad": true,
-  "kalimat-bayyinat": true,
-  "ma-yafcalhu-atibba": true,
-  "masbuk-dhahab": true,
-  "qalaid-marjan": true,
-  "qawl-macruf": true,
-  "rafc-shubha": true,
-  "shahada-zakiyya": true,
-  "tahqiq-rujhan": true,
-  "radd-rawafid": true,
-  "nayl-ibtihaj": true,
-  "tarikh-nur-safir": true,
-  "bahja-mahafil": true,
-  "azhar-riyad": true,
-  "sharh-cala-mukhtasar-khalil": true,
-  "muhadarat-fi-lugha": true,
-  "zahr-akam": true,
-  ishaca: true,
-  "samt-nujum": true,
-  "khulasat-athar": true,
-  "nafahat-rayhana": true,
-  "ithaf-fudala-bashar": true,
-  "ghayth-nafc": true,
-  "tanbih-ghafilin-3": true,
-  "anwar-rabic": true,
-  "sulafa-casr": true,
-  "badr-tamam": true,
-  "bayan-wa-tacrif": true,
-  "tafrij-kurab": true,
-  "sharh-cala-mawahib": true,
-  "sharh-cala-muwatta": true,
-  "fawakih-cadida": true,
-  "mashyakha-29": true,
-  "ruh-bayan": true,
-  "nayl-maarib": true,
-  "fath-wadud": true,
-  "hashiya-cala-ibn-majah": true,
-  "hashiya-cala-nasai": true,
-  "hashiya-cala-sahih-bukhari": true,
-  "tactir-anam": true,
-  "jidd-hathith": true,
-  "tartib-culum": true,
-  "tarikh-yaman": true,
-  "fatawa-khalili": true,
-  "fawaid-jalila": true,
-  ziyada: true,
-  "yawmiyyat-shamiyya": true,
-  "bariqa-mahmudiyya": true,
-  "qitac-jidal": true,
-  "kashf-khafa": true,
-  "fath-ghufur": true,
-  "diwan-islam": true,
-  "sharh-kifaya": true,
-  muhayya: true,
-  "fahrasa-2": true,
-  "tadhyil-sulafat-casr": true,
-  "mutamar-najaf": true,
-  "nafha-miskiyya-1": true,
-  "hawadith-dimashq": true,
-  "ciqd-jayyid": true,
-  "fawz-kabir": true,
-  "hujjat-allah-baligha": true,
-  "insaf-fi-bayan": true,
-  "iclam-multazam": true,
-  "iftiraq-umma": true,
-  "insaf-fi-haqiqat-awliya": true,
-  "irshad-nuqqad": true,
-  "isbal-matar": true,
-  "istifa-aqwal": true,
-  "rafc-astar": true,
-  "risala-latifa": true,
-  "subul-salam": true,
-  "tahbir-1": true,
-  tanwir: true,
-  "tathir-ictiqad": true,
-  "tawdih-afkar": true,
-  "thamarat-nazar": true,
-  "usul-fiqh-1": true,
-  "hulul-tacab": true,
-  "buhur-zakhira": true,
-  "durra-mudiyya-3": true,
-  "ghidha-albab": true,
-  "kashf-litham-1": true,
-  "lawaih-anwar": true,
-  "hashiyat-cadawi": true,
-  "rawd-nadi": true,
-  "bidayat-cabid": true,
-  "fawaid-murdiyya": true,
-  "kashf-mukhaddarat": true,
-  "tuhfa-2": true,
-  carus: true,
-  "hashiya-1": true,
-  "bulghat-arib": true,
-  "ciqd-juman-1": true,
-  "hikmat-ishraq": true,
-  "idah-madarik": true,
-  "adab-mashy-ila-salat": true,
-  "ahadith-fi-fitan": true,
-  "ahkam-salat": true,
-  "ahkam-tamanni-mawt": true,
-  "arbac-qawacid-taduru-ahkam-calayha": true,
-  "bacd-fawaid-sulh-hudaybiyya": true,
-  "fadail-quran-9": true,
-  "fadl-islam": true,
-  "fatawa-5": true,
-  "hadith-78": true,
-  "jawahir-mudiyya": true,
-  "kabair-2": true,
-  "kashf-shubuhat": true,
-  "khutab-minbariyya": true,
-  "kitab-tahara": true,
-  "mabhath-ijtihad-wa-khilaf": true,
-  "macna-la-ilaha-illa-llah-1": true,
-  "majmucat-rasail-fi-tawhid": true,
-  "mansik-hajj-1": true,
-  "masail-jahiliyya": true,
-  "masail-lakhkhasaha-ibn-cabd-wahhab": true,
-  "mufid-mustafid": true,
-  "mukhtasar-insaf": true,
-  "mukhtasar-sira": true,
-  "mukhtasar-tafsir-mucawwidhatayn": true,
-  "mukhtasar-tafsir-surat-anfal": true,
-  "mukhtasar-zad-macad": true,
-  "qawacid-arbaca": true,
-  "rasail-shakhsiyya": true,
-  "risala-fi-radd-cala-rafida": true,
-  "risala-mufida": true,
-  "shurut-salat": true,
-  "tafsir-ayat-min-quran": true,
-} as Record<string, boolean>;
-
-const ComingSoonAlert = async () => {
-  const t = useTranslations("reader");
-  const locale = useLocale() as AppLocale;
-
-  return (
-    <SidebarContainer>
-      <Alert
-        dir={getLocaleDirection(locale)}
-        className="bg-transparent font-sans"
-      >
-        <AlertTitle>{t("coming-soon.title")}</AlertTitle>
-        <AlertDescription>{t("coming-soon.message")}</AlertDescription>
-      </Alert>
-    </SidebarContainer>
-  );
-};
-
-const Placeholder = () => (
-  <div className="flex h-[200px] w-full items-center justify-center">
-    <Spinner className="size-5" />
-  </div>
-);
-
-const AITab = dynamic(() => import("./ai-tab"), { loading: Placeholder });
-const ContentTab = dynamic(() => import("./content-tab"), {
-  loading: Placeholder,
-});
-const SearchTab = dynamic(() => import("./search-tab"), {
-  loading: Placeholder,
-});
-
-type TabId = (typeof tabs)[number]["id"];
-const tabIdToComponent: Record<TabId, any> = {
-  ai: AITab,
-  search: SearchTab,
-  content: ContentTab,
-};
+import type { TabProps } from "./sidebar/tabs";
+import { ComingSoonAlert } from "./coming-soon-alert";
+import { tabIdToComponent } from "./sidebar/tabs-content";
 
 export const TabContent = ({
   bookSlug,
@@ -3329,9 +12,9 @@ export const TabContent = ({
 }: TabProps & {
   tabId: keyof typeof tabIdToComponent;
 }) => {
-  const isSupported = supportedBooks[bookSlug];
+  const isSupported = bookResponse.book.flags.aiSupported;
 
-  if (!isSupported) {
+  if (!isSupported && (tabId === "ai" || tabId === "search")) {
     return <ComingSoonAlert />;
   }
