+++ src/types/index.ts
@@ -1,6 +1,10 @@
 declare global {
   namespace PrismaJson {
     type BookVersion = { source: "openiti" | "turath"; value: string };
+
+    type BookFlags = {
+      aiSupported?: boolean;
+    };
   }
 }
